% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACS_ADS_DE.R
\name{ACS_ADS_DE_plot}
\alias{ACS_ADS_DE_plot}
\title{DE evidence visualization with c-scores and d-scores}
\usage{
ACS_ADS_DE_plot(
  mcmc.merge.list,
  ACS_ADS_pathway,
  dataset.names,
  cluster.lb = NULL,
  select.pathway.list,
  highlight.pathway.index = NULL,
  plot.path = NULL
)
}
\arguments{
\item{mcmc.merge.list:}{a list of merged MCMC output matrices.}

\item{ACS_ADS_pathway:}{a list of four data frames: pathway specific c-scores, d-scores
and their permuted p-value (rows are pathways and columns are studies).}

\item{dataset.names:}{a vector of dataset names.}

\item{cluster.lb:}{a vector of cluster label named by pathway names. optional.}

\item{select.pathway.list:}{a list of selected pathways (containing gene components).}

\item{highlight.pathway.index:}{a numeric vector indicating which pathways in the select.pathway.list need to be highlighted.}
}
\value{
A collection of DE evidence plots for each data pair. In each individual panel, each dot is a pathway sized by p(c-score) (orange in upper right)
or p(d-score) (blue in lower left).X and y axes are the average absolute DE posterior probabilities.
}
\description{
The \code{ACS_ADS_DE_plot} is function to visualize DE evidence and the significance of c-scores and d-scores in one plot. X and y axes represent the average
DE posterior probabilities in each study pair, size of dots represent the significance of c-scores (upper right) or d-scores (lower left).
}
\examples{
\dontrun{
#mcmc.merge.list from the merge step (see the example in function 'merge')
#select.pathway.list from the pathSelect step (see the example in function 'pathSelect')
#ACS_ADS_pathway from the multi_ACS_ADS_pathway step (see the example in 'multi_ACS_ADS_pathway')
#highlight and label the first three pathways in selec.pathway.list:
highlight.pathway.index = c(1,2,3)
ACS_ADS_DE_plot = function(mcmc.merge.list,ACS_ADS_pathway,dataset.names,
                           cluster.lb=NULL,select.pathway.list,
                           highlight.pathway.index=highlight.pathway.index,plot.path=NULL)
}
}
