% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addnumbers.R
\name{addNumbers}
\alias{addNumbers}
\title{this is the description}
\usage{
addNumbers(first, second)
}
\arguments{
\item{first}{this is the first number}

\item{second}{this is the second number}
}
\value{
the return is the sum of the two numbers
}
\description{
this is the description
}
