% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/onload.R
\name{assign_pkg_globals}
\alias{assign_pkg_globals}
\title{Assign variables to package-global environment.
Detach (recursively), assign, attach.
pkg_globals should be at search path position 2.}
\usage{
assign_pkg_globals(varName, varValue)
}
\arguments{
\item{varName}{Name of variable as character}

\item{varValue}{Value of character}
}
\description{
Assign variables to package-global environment.
Detach (recursively), assign, attach.
pkg_globals should be at search path position 2.
}

