% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{bayes}
\alias{bayes}
\title{Threshold-free Bayesian differential analysis}
\usage{
bayes(pData, seed = 12345)
}
\arguments{
\item{pData:}{a matrix consist of two columns where first being the p-values and second being the log fold changes.}

\item{seed:}{seed number.}
}
\value{
an MCMC output matrix of signed DE indicator (input for
congruence analysis)
}
\description{
The \code{bayes} is function to run Bayesian differential analysis for a single study.
}
\examples{
\dontrun{
data(hb)
summaryDE = indDE(data=data,group=as.factor(group),data.type="microarray",
                  case.label="2", ctrl.label="1")
hb_pData = summaryDE[,c(3,1)]
hb_MCMCout = bayes(hb_pData, seed=12345)
}
}
