% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BingxueF.R
\name{BingxueF}
\alias{BingxueF}
\title{Bingxue}
\usage{
BingxueF(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
the square of x
}
\description{
Practice creating R packages
}
\examples{
BingxueF(3)

}
\author{
Bingxue Wu
}
