% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpval.R
\name{bootpval}
\alias{bootpval}
\title{Bootstrap P-values}
\usage{
bootpval(x, conf.level = 0.95, iter = 3000, mu0 = 0, test = "two")
}
\arguments{
\item{x}{vector containing a sample}

\item{conf.level}{confidence level for the confidence interval}

\item{iter}{amount of iterations for the bootstrap}

\item{mu0}{mean of the null hypothesis}

\item{test}{type of test}
}
\value{
a histogram showing the T-Distribution and p-value areas
}
\description{
Bootstrap P-values
}
\examples{
\dontrun{
bootpval(x=x1,mu0=22)
}
}
