% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cali.plot.R
\name{cali.plot}
\alias{cali.plot}
\title{Title}
\usage{
cali.plot(
  variable,
  time = "OS",
  status = "EVENT",
  data = df,
  t1 = 1095,
  t2 = 1825
)
}
\arguments{
\item{variable}{character variable which is factor}

\item{time}{survival time}

\item{status}{survival status}

\item{data}{data.frame style contain variable which is factor}

\item{t1}{3 year (day)}

\item{t2}{5 year (day)}
}
\value{

}
\description{
Title
}
\examples{
variable <- names(df)[2:4]
cali.plot(variable,time="OS",status="EVENT",data=df,t1=1095,t2=1825)

}
