% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_videos.R
\name{category_to_id}
\alias{category_to_id}
\title{Converts content category to YouTube id}
\usage{
category_to_id(key, category, region = "US")
}
\arguments{
\item{region}{String containing content category of intrest.}
}
\value{
A string of a single category id
}
\description{
Converts content category to YouTube id
}
