% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_titles_word_cloud.R
\name{clean_titles}
\alias{clean_titles}
\title{Cleans a list of YouTube video titles}
\usage{
clean_titles(titles)
}
\arguments{
\item{A}{list of YouTube videos titles to be cleaned}
}
\value{
A cleaned list of YouTube titles
}
\description{
Cleans a list of YouTube video titles
}
