% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_videos.R
\name{clean_top_videos}
\alias{clean_top_videos}
\title{Cleans top vidoes data after scraping}
\usage{
clean_top_videos(data)
}
\arguments{
\item{data}{Nested data table pulled from YouTube API}
}
\value{
A clean data frame of the top YouTube videos
}
\description{
Cleans top vidoes data after scraping
}
