% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_plot.R
\name{cor_plot}
\alias{cor_plot}
\title{cor_plot}
\usage{
cor_plot(cor.out, results = length(cor.out$cor), graph = "bar")
}
\arguments{
\item{cor.out}{a cor object, as created by the cor_fun().}

\item{results}{sets the number of total values that will be plotted. By default it is set to the number of cor values. Also, the maximum number of values that can be plotted is 1000.}

\item{graph}{sets the graphic type. Can take the values 'line' or 'bar'. By default is is set to 'bar'.}
}
\value{
a plot of cross-correlated values is returned.
}
\description{
Plots the cor_fun output object using ggplot2. It is designed to only be used in conjunction with the output of the cor_fun() function ('cor.out'). The dotted red lines (shown only for samples), are 95% confidence intervals calculated from the original correlation.
}
