% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_pop.R
\name{cor_pop}
\alias{cor_pop}
\title{cor_pop}
\usage{
cor_pop(x, y)
}
\arguments{
\item{x}{a vector of numeric values}

\item{y}{a value of numeric values}
}
\value{
the population level correlation of the inputs is returned.
}
\description{
Calculates the correlation of two vectors using n as the denominator (suitable for populations) rather than n-1 (suitable for samples).
}
