% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_sample.R
\name{cor_sample}
\alias{cor_sample}
\title{cor_sample}
\usage{
cor_sample(x, y)
}
\arguments{
\item{x}{a vector of numeric values}

\item{y}{a value of numeric values}
}
\value{
the sample level correlation of the inputs is returned.
}
\description{
Calculates the correlation of two vectors using n-1 as the denominator (suitable for samples), rather than n (suitable for populations).
}
