% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_scatter.R
\name{cor_scatter}
\alias{cor_scatter}
\title{cor_scatter}
\usage{
cor_scatter(x, y = NULL, lags, period = "", x.lab = "x", y.lab = "y")
}
\arguments{
\item{x}{a vector of numeric values}

\item{y}{a vector of numeric values}

\item{lags}{a list containing the values to be lagged}

\item{period}{the time period over which values are lagged, e.g. days or hours}

\item{x.lab}{the label for the x axes}

\item{y.lab}{the label for the y axes}
}
\value{
a lattice of scatter plots.
}
\description{
Plots separate scatter plots in a lattice for several lag values
}
