% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_pop.R
\name{cov_pop}
\alias{cov_pop}
\title{cov_pop}
\usage{
cov_pop(x, y)
}
\arguments{
\item{x}{a vector of numeric values}

\item{y}{a vector of numeric values}
}
\value{
the covariance of the inputs is returned.
}
\description{
Calculates the covariance of two vectors using n as the denominator (suitable for populations) rather than n-1 (suitable for samples).
}
