% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grDevices.R
\name{createAllDevPlots}
\alias{createAllDevPlots}
\title{create plots on all devices}
\usage{
createAllDevPlots(plotFun, title, pdfHeight = NULL, pngHeight = NULL,
  pngWidth = NULL, pngRes = NULL, ...)
}
\arguments{
\item{plotFun}{function creating a plot}

\item{title}{Title for the plot}

\item{pdfHeight}{The PDF plot height in inches}

\item{pngHeight}{The PNG plot height in inches}

\item{pngWidth}{The PNG plot width in inches}

\item{pngRes}{The PNG resolution in dots per inch}

\item{...}{Additional parameters for plotFun}
}
\description{
create plots on all devices
}

