% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPackage.R
\name{draw_comic}
\alias{draw_comic}
\title{Plot an xkcd comic strip}
\usage{
draw_comic(x)
}
\arguments{
\item{x}{A list object as retruned from \code{\link[=xkcd]{xkcd()}}}
}
\value{
A rastergrob grob. See"Details" section of the \code{\link[grid:grid.grob]{grid::grob()}} for more information.
}
\description{
'draw_comic()' display and xkcd comic strip  as aplot in a graphics device (e.g., the RStudio window).
}
\examples{
dino_time<- xkcd(15)
str(dino_time)
}
