% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyfreq.R
\name{dummyfreq}
\alias{dummyfreq}
\title{Create frequency table for multi-option dummy variable}
\usage{
dummyfreq(data, prefix, label_list, title)
}
\arguments{
\item{data}{your data frame}

\item{prefix}{prefix that identifies all dummy variables in one category}

\item{label_list}{optional: list of labels for each dummy level}

\item{title}{optional: title of Table 1}
}
\value{
table 1 displaying n (%) of each dummy variable level
}
\description{
Create frequency table for multi-option dummy variable
}
\examples{
library(dplyr)
library(table1)
category_names <- c("Health", "Education", "Transportation", "Social services", "Safety", "Diversity & inclusion", "Housing")
dummyfreq(data=randomsurvey, prefix="Aspect", label_list=category_names, title="Community aspects")
}
\author{
Katherine Lu
}
