% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyvar.R
\name{dummyvar}
\alias{dummyvar}
\title{Combine dummy variables with exclusive levels}
\usage{
dummyvar(data, prefix, label_list, varname)
}
\arguments{
\item{data}{your data frame}

\item{prefix}{prefix that identifies all dummy variables in one category}

\item{label_list}{optional: list of labels for each dummy level}

\item{varname}{optional: label for combined variable}
}
\value{
data frame with new combined variable
}
\description{
Combine dummy variables with exclusive levels
}
\examples{
library(dplyr)
library(table1)
area_labels <- c("Urban", "Suburban", "Rural")
data2 <- dummyvar(data=randomsurvey, prefix="AreaType", label_list=area_labels, varname="Area Types")
table1(~ combined, data=data2) #see new variable in a table1
}
\author{
Katherine Lu
}
