% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorPlots_fun.R
\name{error_bars}
\alias{error_bars}
\title{error_bars}
\usage{
error_bars(df, x, s = NULL)
}
\arguments{
\item{df}{name of the dataframe that you want to analyse.}

\item{x}{variable within the dataframe you want to bin over. Should be provided in the form 'df$x'}

\item{s}{sequence of values (generated by seq(x,y,z)) - optional input: if not provided it will be calculated automatically.}
}
\value{
an 'error_bars.out' object which contains (i) a dataframe of useful summary variables, and (b) a dot plot showing the error bars.
}
\description{
analysis of error bars for a variable
}
