% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Function}
\usage{
euclidean(a, b)
}
\arguments{
\item{a, b}{are the integers}
}
\value{
Greatest Common divisor of these two numbers
}
\description{
Euclidean Function
}
