% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{plot the accidents of a particular state in a particular year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the number of a state}

\item{year}{a range of years}
}
\value{
return a graph of the location of the accidents of a particular state in a particular year
}
\description{
plot the accidents of a particular state in a particular year
}
