% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read CSV files}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a string that indicates the path/name of the file}
}
\value{
a dataframe tbl from the csv file
}
\description{
Read CSV files
}
\details{
will return error if the no file of specified name exists
}
