% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{read files with name in the range years and return a dataframe with column month and year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a range of years}
}
\value{
a dataframe of month and year where the file contains such name exists
}
\description{
read files with name in the range years and return a dataframe with column month and year
}
\details{
warning will return if the file of the year does not exist
import dplyr::mutate, dplyr::select
}
