% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{summarize the number of year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a range of years}
}
\value{
return a data frame which contains a column of the number of columns with the same year and month
import dplyr::bind_rows, dplyr::group_by, dplyr::summarize, tidyr::spread
}
\description{
summarize the number of year and month
}
