% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.alpha.perplexity.R
\name{fit.alpha.perplexity}
\alias{fit.alpha.perplexity}
\title{fit.alpha.perplexity}
\usage{
fit.alpha.perplexity(dtm, folds, alpha.values, k, beta, control.test)
}
\arguments{
\item{dtm}{Document-term matrix. Constructed using the DocumentTermMatrix() command from the tm package}

\item{folds}{Integer. The number of folds to make training and testing sets; recommended values are '5' and '10' - note that higher values considerably increase the time that model fitting takes}

\item{alpha.values}{Numeric vector. Values to test alpha for. A good starting point is c(0.001, 0.01, 0.1, 1)}

\item{k}{Integer. Optional parameter: the value of k used in the LDA model. By default k is set to 10}

\item{beta}{Numeric. Optional parameter: the value of beta used in the LDA model. By default beta is set to 0.1}

\item{control.test}{List. Optional parameter: the LDA control list used in the LDA model. It is strongly recommended not to use this parameter unless you have good reason. Default settings are: nstart = 5, best = T, burnin = 1000, iter = 2000, thin = 500}
}
\value{
Dataframe of perplexity for the alpha.values, calculated for the number of stipulated folds
}
\description{
Calculates perplexity for values of alpha on an LDA topic model using the topicmodels package, split into training and testing sets using k-folds
}
