% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{foo}
\alias{foo}
\title{For Orientation Only}
\usage{
foo(msg = "Are you sleeping?")
}
\arguments{
\item{msg}{a message to \code{cat} to the console.}
}
\description{
This function does nothing useful, but it could!
}
\examples{
foo("No, I am not!")
}
