% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_titles_word_cloud.R
\name{format_titles}
\alias{format_titles}
\title{Formats a list of YouTube video titles to be}
\usage{
format_titles(clean_titles)
}
\arguments{
\item{A}{list of YouTube videos titles to be formatted for a word cloud}
}
\value{
A formatted titles dataframe
}
\description{
Formats a list of YouTube video titles to be
}
