% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_videos.R
\name{get_category_list}
\alias{get_category_list}
\title{Gets list of YouTube content categories}
\usage{
get_category_list(key, region = "US")
}
\arguments{
\item{key}{Your YouTube API key}

\item{region}{Region. Default is "US". Run get_region_list() to see list of YouTube regions}
}
\value{
A data frame of YouTube content categories and their ids
}
\description{
Gets list of YouTube content categories
}
