% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_cloud.R
\name{get_comments}
\alias{get_comments}
\title{Access the youtube API to get a list of comments from a video.}
\usage{
get_comments(api_key, video_id)
}
\arguments{
\item{api_key}{The Youtube API key}

\item{video_id}{ID of youtube video}
}
\value{
A list of comments
}
\description{
Access the youtube API to get a list of comments from a video.
}
