% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/onload.R
\name{get_pkg_globals}
\alias{get_pkg_globals}
\title{Get variables from package-global environment.}
\usage{
get_pkg_globals(varName)
}
\arguments{
\item{varName}{Name of variable as character}
}
\description{
Get variables from package-global environment.
}

