% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_videos.R
\name{get_region_list}
\alias{get_region_list}
\title{Gets list of YouTube content regions}
\usage{
get_region_list(key)
}
\arguments{
\item{key}{Your YouTube API key}
}
\value{
A data frame of YouTube regions and their ids
}
\description{
Gets list of YouTube content regions
}
