% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_titles_word_cloud.R
\name{get_tidy_titles}
\alias{get_tidy_titles}
\title{Creates a list of cleaned list of YouTube video titles for word cloud and sentiment analysis}
\usage{
get_tidy_titles(titles)
}
\arguments{
\item{A}{list of YouTube videos titles to be cleaned}
}
\value{
A cleaned list of YouTube titles
}
\description{
Creates a list of cleaned list of YouTube video titles for word cloud and sentiment analysis
}
