% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_titles_word_cloud.R
\name{get_top_titles}
\alias{get_top_titles}
\title{Returns a list of 200 titles from the most popular YouTube video pages}
\usage{
get_top_titles(api_key)
}
\arguments{
\item{api_key}{a user's API key}
}
\value{
A list of 200 titles from the most popular YouTube video pages
}
\description{
Returns a list of 200 titles from the most popular YouTube video pages
}
