% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_videos.R
\name{get_top_videos}
\alias{get_top_videos}
\title{Makes data frame of top YouTube videos}
\usage{
get_top_videos(key, region = "US", category = 0, n = 10, simple = T)
}
\arguments{
\item{key}{Your YouTube API key}

\item{region}{Region. Default is "US". Run get_region_list() to see list of YouTube cntent regions.}

\item{category}{Video category. Default is 0 (no category, top videos overall). Run get_category_list() to see list of YouTube content categories.}

\item{n}{Number of results returned. Default is 10. Can be between 1 and 50.}

\item{simple}{Returns a simplified data frame if TRUE (default). Returns all data from API if FALSE.}
}
\value{
A data frame of the top trending videos on YouTube
}
\description{
Makes data frame of top YouTube videos
}
