% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM.R
\name{GMM}
\alias{GMM}
\title{Perform Gaussian Mixture Model on a data matrix.}
\usage{
GMM(data.mat, K)
}
\arguments{
\item{K}{The number of clusters}

\item{data.matrix}{A numeric data matrix (nxp)}
}
\value{
A numeric value of log.likely error
}
\description{
This function computes the negative log likely on a data matrix using the
Gaussian Mixture Model
}
\examples{
data.mat <- as.matrix(iris[, c("Petal.Length", "Petal.Width")])
K <- 3
log.likely <- GMM(data.mat, K)

}
