% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_videos.R
\name{graph_top_videos_category}
\alias{graph_top_videos_category}
\title{Creates a bar graphs of the most currently popular YouTube Videos by content category}
\usage{
graph_top_videos_category(key, region = "US", n = "10")
}
\arguments{
\item{key}{Your YouTube API key}

\item{region}{Region. Default is "US". Run get_region_list() to see list of YouTube cntent regions.}

\item{n}{Number of top videos to graph. Default is 10. Can be between 1 and 200.}
}
\value{
A bar graph of the most currently popular YouTube Videos by content category
}
\description{
Creates a bar graphs of the most currently popular YouTube Videos by content category
}
