% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ha.group.R
\docType{data}
\name{ha.group}
\alias{ha.group}
\title{Sample data of human ARDS}
\format{
a vector indicating the disease status of samples in ha.
1 indicates control and 2 indicates disease.
}
\usage{
ha.group
}
\description{
Sample data of human ARDS
}
\keyword{datasets}
