% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ha.R
\docType{data}
\name{ha}
\alias{ha}
\title{Gene expression data of human ARDS}
\format{
a matrix where rows are genes and columns are samples
}
\usage{
ha
}
\description{
The data has been preprocessed.
}
\keyword{datasets}
