% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb.group.R
\docType{data}
\name{hb.group}
\alias{hb.group}
\title{Sample data of human burn}
\format{
a vector indicating the disease status of samples in hb.
1 indicates control and 2 indicates disease.
}
\usage{
hb.group
}
\description{
Sample data of human burn
}
\keyword{datasets}
