% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb.R
\docType{data}
\name{hb}
\alias{hb}
\title{Gene expression data of human burn}
\format{
a matrix where rows are genes and columns are samples
}
\usage{
hb
}
\description{
The data has been preprocessed.
}
\keyword{datasets}
