% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{hello}
\alias{hello}
\title{print hello with name}
\usage{
hello(name)
}
\arguments{
\item{name}{name of person}
}
\value{
returns hello name
}
\description{
print hello with name
}
\examples{
hello("sunny")
}
