% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi.group.R
\docType{data}
\name{hi.group}
\alias{hi.group}
\title{Sample data of human infection}
\format{
a vector indicating the disease status of samples in hi.
1 indicates control and 2 indicates disease.
}
\usage{
hi.group
}
\description{
Sample data of human infection
}
\keyword{datasets}
