% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hi}
\alias{hi}
\title{hi - a simple R object}
\format{
an R \cite{"character"}
}
\usage{
hi
}
\description{
Save R objects with save and access their documentation with the help
commend. For example, help("ozone",package="plyr").
}
\references{
\url{https://github.com/hadley/plyr/blob/master/R/data.r}
}
\keyword{datasets}
