% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl.group.R
\docType{data}
\name{hl.group}
\alias{hl.group}
\title{Sample data of human LPS}
\format{
a vector indicating the disease status of samples in hl.
1 indicates control and 2 indicates disease.
}
\usage{
hl.group
}
\description{
Sample data of human LPS
}
\keyword{datasets}
