% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl.R
\docType{data}
\name{hl}
\alias{hl}
\title{Gene expression data of human LPD}
\format{
a matrix where rows are genes and columns are samples
}
\usage{
hl
}
\description{
The data has been preprocessed.
}
\keyword{datasets}
