% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs.group.R
\docType{data}
\name{hs.group}
\alias{hs.group}
\title{Sample data of human sepsis}
\format{
a vector indicating the disease status of samples in hs.
1 indicates control and 2 indicates disease.
}
\usage{
hs.group
}
\description{
Sample data of human sepsis
}
\keyword{datasets}
