% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs.R
\docType{data}
\name{hs}
\alias{hs}
\title{Gene expression data of human sepsis}
\format{
a matrix where rows are genes and columns are samples
}
\usage{
hs
}
\description{
The data has been preprocessed.
}
\keyword{datasets}
