% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht.group.R
\docType{data}
\name{ht.group}
\alias{ht.group}
\title{Sample data of human trauma}
\format{
a vector indicating the disease status of samples in ht.
1 indicates control and 2 indicates disease.
}
\usage{
ht.group
}
\description{
Sample data of human trauma
}
\keyword{datasets}
