% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht.R
\docType{data}
\name{ht}
\alias{ht}
\title{Gene expression data of human trauma}
\format{
a matrix where rows are genes and columns are samples
}
\usage{
ht
}
\description{
The data has been preprocessed.
}
\keyword{datasets}
