% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indDE.R
\name{indDE}
\alias{indDE}
\title{Differential expression analysis for an individual study.}
\usage{
indDE(data, group, data.type, case.label, ctrl.label)
}
\arguments{
\item{data:}{the gene expression matrix}

\item{group:}{the group label.}

\item{data.type:}{"microarray" or "RNAseq". LIMMA is performed for microarray data
while DEseq2 is performed for RNAseq data.}

\item{case.label:}{the label name for the case group.}

\item{control.label:}{the label name for the control group.}
}
\value{
a summary data frame of log fold change, lfc standard error,
p-value and q-value.
}
\description{
The \code{indDE} is function to perform differential expression
analysis for an individual study.
}
\examples{
\dontrun{
data(hb)
summaryDE = indDE(data=data,group=as.factor(group),data.type="microarray",
                   case.label="2", ctrl.label="1")
}
}
