% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{installed_and_attached}
\alias{installed_and_attached}
\title{Installed and Attached when Called}
\usage{
installed_and_attached()
}
\description{
Installed with the package via \code{Import} in \code{DESCRIPTION} and
package search path specified with \code{::} when function called.
}
