% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{installed_and_dependent}
\alias{installed_and_dependent}
\title{Installed and Dependent but not Attached when Called}
\usage{
installed_and_dependent()
}
\description{
Installed with the package via \code{Import} in \code{DESCRIPTION} and
package search path specified with \code{::} when function called.
}
\note{
Does not error out. The user can also use this package directly.
For example, try \code{mtcars  \%>\% select(disp)} without a \code{library} call.

It doesn't find the function in the package namespace, but it works
because it eventually looks up the function in the user's search path
(in which \code{dplyr} is attached). But you would rather use \verb{@import} here
(see \link{installed_and_imported}) unless you want the \emph{user} to
have the ability to use \code{dplyr} without explicitly attaching.
}
