% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{installed_and_imported}
\alias{installed_and_imported}
\title{Installed and Imported to \code{NAMESPACE}}
\usage{
installed_and_imported()
}
\description{
Installed with the package via \code{Import} in \code{DESCRIPTION} and
packaged added to the \code{NAMESPACE} by \verb{@import}. Function from
package is called without \code{::}.
}
\note{
Expect this to work. But the user can't use the functions without
attaching the package (in this case \code{tidyr}). Try \code{drop_na(mtcars)}.
}
