% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{installed_not_attached}
\alias{installed_not_attached}
\title{Installed but Not Attached}
\usage{
installed_not_attached()
}
\description{
Installed with the package via \code{Import} in \code{DESCRIPTION} but the
function is called from the package without attaching the package.
}
\note{
Expect it to error out.
}
