% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG_module_topology_plot.R
\name{KEGG_module_topology_plot}
\alias{KEGG_module_topology_plot}
\title{KEGG topology plot with highlighted module nodes.}
\usage{
KEGG_module_topology_plot(
  res_KEGG_module,
  which_to_draw = "all",
  filePath = getwd()
)
}
\arguments{
\item{res_KEGG_module:}{a result list from function \code{KEGG_module}.}

\item{which_to_draw:}{either a numeric vector indicating module sizes of interest to highlight or "all"
which will show all module size scenarios in the res_KEGG_module provided.}

\item{filePath:}{the path to save the elbow plot. Default is the current working directory.}
}
\value{
KEGG pathway topology plots with different module highlighted will be saved as .png
files in the filePath provided.
}
\description{
The \code{KEGG_module_topology_plot} is a function to highlght module location on KEGG pathway
topology.
}
\examples{
\dontrun{
#res_KEGG_module from the KEGG_module step (see the example in function 'KEGG_module')
res = KEGG_module_topology_plot(res_KEGG_module,which_to_draw = c(4,8,9))
}
}
