% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{methods}
\name{koef.reg}
\alias{koef.reg}
\title{Get Regression Coefficients}
\usage{
koef.reg(model)
}
\arguments{
\item{model}{An object of class \code{RLB}.}
}
\value{
A numeric vector representing the estimated coefficients.
}
\description{
Extracts the estimated regression coefficients from an RLB object.
}
