% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{methods}
\name{koef.reg}
\alias{koef.reg}
\title{Ekstraktor untuk Koefisien Regresi dari Objek RLB}
\usage{
koef.reg(model)
}
\arguments{
\item{model}{Objek dari kelas \code{RLB}.}
}
\value{
\code{numeric} Vektor koefisien regresi.
}
\description{
Mengembalikan vektor koefisien regresi dari objek kelas RLB.
}
