% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{koef.reg}
\alias{koef.reg}
\title{Aksesor untuk Koefisien Regresi}
\usage{
koef.reg(model)
}
\arguments{
\item{model}{Objek kelas RLB.}
}
\value{
Mengembalikan slot koef.
}
\description{
Mengambil slot koef dari objek RLB.
}
