% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/length_function.R
\name{length_function}
\alias{length_function}
\title{The Length Function}
\usage{
length_function(sales.balance = 1500, Option = "LowPayment", gamma = -1.8,
  alpha = 0.0018, c = 1.2, delta = 136.8)
}
\arguments{
\item{Option}{"LowPayment" or "LowCost" should be entered to return the correspoding length}

\item{gamma}{Do not adjust this parameter in production. Defaults to -1.8}

\item{alpha}{Do not adjust this parameter in production. Defaults to -0.0018}

\item{c}{Do not adjust this parameter in production. Defaults to 1.2}

\item{delta}{Do not adjust this parameter in production. Defaults to 136.8}

\item{sales.balanace}{The sales price balance (sales price-down payent) for the item on the contract. Defaults to 1500.}
}
\description{
This function returns the length of a lease determined by the sales price balance.  The function returns the length in weeks.
}
\examples{
length_function(sales.balance=1500,Option=c("LowCost","LowPayment"))
}
\keyword{Length}

