% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab4.R
\name{lmean}
\alias{lmean}
\title{lmean}
\usage{
lmean(list_of_vectors)
}
\arguments{
\item{list_of_vectors}{\code{numeric} vector}
}
\description{
lmean
}
\details{
Calculates mean of every list element
}
\examples{
# random comment
lmean(list(c(12,3,4,5,6,7,4),c(12,3,4,5,6,7)))
}
\seealso{
\code{\link[base]{mean}}
}
\author{
Sebastian Sowik
}
