% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.group.R
\docType{data}
\name{ma.group}
\alias{ma.group}
\title{Sample data of mouse ARDS}
\format{
a vector indicating the disease status of samples in ma.
1 indicates control and 2 indicates disease.
}
\usage{
ma.group
}
\description{
Sample data of mouse ARDS
}
\keyword{datasets}
