% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.R
\docType{data}
\name{ma}
\alias{ma}
\title{Gene expression data of mouse ARDS}
\format{
a matrix where rows are genes and columns are samples
}
\usage{
ma
}
\description{
The data has been preprocessed.
}
\keyword{datasets}
