% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_cloud.R
\name{make_cloud}
\alias{make_cloud}
\title{Create a word cloud from a list of comments.}
\usage{
make_cloud(api_key, video_id)
}
\arguments{
\item{api_key}{The Youtube API key}

\item{video_id}{ID of youtube video}
}
\value{
A word cloud
}
\description{
Create a word cloud from a list of comments.
}
