% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make a file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a numeric used for name}
}
\value{
a string as a name for bz2 file that takes in year
}
\description{
make a file name
}
\details{
error if year is not a numeric
}
