% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_titles_word_cloud.R
\name{make_top_title_word_cloud}
\alias{make_top_title_word_cloud}
\title{Creates and returns a word cloud of 200 of the titles on YouTube's most popular videos}
\usage{
make_top_title_word_cloud(api_key, size, color, shape)
}
\arguments{
\item{api_key}{A user's API key}

\item{size}{The size of the title word cloud to be returned}

\item{color}{The color of the title word cloud to be returned}

\item{shape}{The shape fo the title word cloud to be returned}
}
\value{
A word cloud using 200 of the titles from YouTube's most popular videos
}
\description{
Creates and returns a word cloud of 200 of the titles on YouTube's most popular videos
}
