% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_pts.R
\name{max_pts}
\alias{max_pts}
\title{Maximum Points Function}
\usage{
max_pts(data, year)
}
\arguments{
\item{year}{Which is your year of interest? Range from 1950 to 2017.}
}
\description{
This function get the oldest player of the input year, if multiple players found, return the one earned the most points
}
\examples{
max_pts()

}
\keyword{points}
