% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb.group.R
\docType{data}
\name{mb.group}
\alias{mb.group}
\title{Sample data of mouse burn}
\format{
a vector indicating the disease status of samples in mb.
1 indicates control and 2 indicates disease.
}
\usage{
mb.group
}
\description{
Sample data of mouse burn
}
\keyword{datasets}
