% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb.R
\docType{data}
\name{mb}
\alias{mb}
\title{Gene expression data of mouse burn}
\format{
a matrix where rows are genes and columns are samples
}
\usage{
mb
}
\description{
The data has been preprocessed.
}
\keyword{datasets}
