% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge}
\alias{merge}
\title{Merge multiple MCMCout matrices by orthologs}
\usage{
merge(mcmc.list, species, ortholog.db = NULL, reference = 1, uniqG = T)
}
\arguments{
\item{mcmc.list:}{a list of MCMC output matrices.}

\item{species:}{a vector specie names of same length as mcmc.list indicating
the species of each study.}

\item{ortholog.db:}{a data.frame/matrix match orthologs between species.
Column names should be consistent with the species input. If not provided,
datasets will be merges without orthologs matching.}

\item{reference:}{the index of the reference MCMC matrix. Merged
list will be named using the rownames of this matrix.}

\item{uniqG:}{TRUE: only keep the gene with greatest posterior DE signal
when multiple matches provided in the orhthologs file. FALSE: keep duplicated
genes when multiple matches exist. default is TRUE.}
}
\value{
an merged list of multiple MCMCout datasets (with same number of
rows and rownames)
}
\description{
The \code{merge} is function to merge multiple MCMCout matrices by
matching genes/orthologs.
}
\examples{
\dontrun{
data(hb)
summaryDE = indDE(data=data,group=as.factor(group),data.type="microarray",
                  case.label="2", ctrl.label="1")
hb_pData = summaryDE[,c(3,1)]
hb_MCMCout = bayes(hb_pData, seed=12345)
data(hs)
summaryDE = indDE(data=data,group=as.factor(group),data.type="microarray",
                  case.label="2", ctrl.label="1")
hs_pData = summaryDE[,c(3,1)]
hs_MCMCout = bayes(hs_pData, seed=12345)
data(ht)
summaryDE <- indDE(data=data,group=as.factor(group),data.type="microarray",
                  case.label="2", ctrl.label="1")
ht_pData <- summaryDE[,c(3,1)]
ht_MCMCout <- bayes(ht_pData, seed=12345)
data(ha)
summaryDE = indDE(data=data,group=as.factor(group),data.type="microarray",
                  case.label="2", ctrl.label="1")
ha_pData = summaryDE[,c(3,1)]
ha_MCMCout = bayes(ha_pData, seed=12345)
data(hi)
summaryDE = indDE(data=data,group=as.factor(group),data.type="microarray",
                  case.label="2", ctrl.label="1")
hi_pData = summaryDE[,c(3,1)]
hi_MCMCout = bayes(hi_pData, seed=12345)
data(hl)
summaryDE = indDE(data=data,group=as.factor(group),data.type="microarray",
                  case.label="2", ctrl.label="1")
hl_pData = summaryDE[,c(3,1)]
hl_MCMCout = bayes(hl_pData, seed=12345)
data(hb)
summaryDE = indDE(data=data,group=as.factor(group),data.type="microarray",
                  case.label="2", ctrl.label="1")
mb_pData = summaryDE[,c(3,1)]
mb_MCMCout = bayes(mb_pData, seed=12345)
data(hs)
summaryDE = indDE(data=data,group=as.factor(group),data.type="microarray",
                  case.label="2", ctrl.label="1")
ms_pData = summaryDE[,c(3,1)]
ms_MCMCout = bayes(ms_pData, seed=12345)
data(ht)
summaryDE <- indDE(data=data,group=as.factor(group),data.type="microarray",
                  case.label="2", ctrl.label="1")
mt_pData <- summaryDE[,c(3,1)]
mt_MCMCout <- bayes(mt_pData, seed=12345)
data(ma)
summaryDE = indDE(data=data,group=as.factor(group),data.type="microarray",
                  case.label="2", ctrl.label="1")
ma_pData = summaryDE[,c(3,1)]
ma_MCMCout = bayes(ma_pData, seed=12345)
data(hi)
summaryDE = indDE(data=data,group=as.factor(group),data.type="microarray",
                  case.label="2", ctrl.label="1")
mi_pData = summaryDE[,c(3,1)]
mi_MCMCout = bayes(mi_pData, seed=12345)
data(ml)
summaryDE = indDE(data=data,group=as.factor(group),data.type="microarray",
                  case.label="2", ctrl.label="1")
ml_pData = summaryDE[,c(3,1)]
ml_MCMCout = bayes(ml_pData, seed=12345)

mcmc.list <- list(hb_MCMCout,hs_MCMCout,ht_MCMCout,
ha_MCMCout,hi_MCMCout,hl_MCMCout,
mb_MCMCout,ms_MCMCout,mt_MCMCout,
ma_MCMCout,mi_MCMCout,ml_MCMCout)
data(hm_orth)
species <- c(rep("human",6), rep("mouse",6))
mcmc.merge.list <- merge(mcmc.list,species = species,
ortholog.db = hm_orth, reference=1,uniqG=T)
}
}
