% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi.group.R
\docType{data}
\name{mi.group}
\alias{mi.group}
\title{Sample data of mouse infection}
\format{
a vector indicating the disease status of samples in mi.
1 indicates control and 2 indicates disease.
}
\usage{
mi.group
}
\description{
Sample data of mouse infection
}
\keyword{datasets}
