% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi.R
\docType{data}
\name{mi}
\alias{mi}
\title{Gene expression data of mouse infection}
\format{
a matrix where rows are genes and columns are samples
}
\usage{
mi
}
\description{
The data has been preprocessed.
}
\keyword{datasets}
