% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.group.R
\docType{data}
\name{ml.group}
\alias{ml.group}
\title{Sample data of mouse LPS}
\format{
a vector indicating the disease status of samples in ml.
1 indicates control and 2 indicates disease.
}
\usage{
ml.group
}
\description{
Sample data of mouse LPS
}
\keyword{datasets}
