% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\docType{data}
\name{ml}
\alias{ml}
\title{Gene expression data of mouse LPS}
\format{
a matrix where rows are genes and columns are samples
}
\usage{
ml
}
\description{
The data has been preprocessed.
}
\keyword{datasets}
