% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.group.R
\docType{data}
\name{ms.group}
\alias{ms.group}
\title{Sample data of mouse sepsis}
\format{
a vector indicating the disease status of samples in ms.
1 indicates control and 2 indicates disease.
}
\usage{
ms.group
}
\description{
Sample data of mouse sepsis
}
\keyword{datasets}
