% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.R
\docType{data}
\name{ms}
\alias{ms}
\title{Gene expression data of mouse sepsis}
\format{
a matrix where rows are genes and columns are samples
}
\usage{
ms
}
\description{
The data has been preprocessed.
}
\keyword{datasets}
