% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt.group.R
\docType{data}
\name{mt.group}
\alias{mt.group}
\title{Sample data of mouse trauma}
\format{
a vector indicating the disease status of samples in mt.
1 indicates control and 2 indicates disease.
}
\usage{
mt.group
}
\description{
Sample data of mouse trauma
}
\keyword{datasets}
