% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt.R
\docType{data}
\name{mt}
\alias{mt}
\title{Gene expression data of mouse trauma}
\format{
a matrix where rows are genes and columns are samples
}
\usage{
mt
}
\description{
The data has been preprocessed.
}
\keyword{datasets}
