% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtx_cor.R
\name{mtx_cor}
\alias{mtx_cor}
\title{A Matrix Correlation Function}
\usage{
mtx_cor(year)
}
\arguments{
\item{year}{which year are you interested in?}
}
\description{
This function will calculate the correlation matrix for all numeric variables in the NBA data for year specified
}
\examples{
mtx_cor(1966)
}
\keyword{mtx}
