% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ACS_ADS_global.R
\name{multi_ACS_ADS_global}
\alias{multi_ACS_ADS_global}
\title{Genome-wide congruence analysis for multiple pairs}
\usage{
multi_ACS_ADS_global(
  mcmc.merge.list,
  dataset.names,
  measure = "Fmeasure",
  B = 100
)
}
\arguments{
\item{mcmc.merge.list:}{a list of merged MCMC output matrices.}

\item{dataset.names:}{a vector of dataset names.}

\item{measure:}{three types of scores to be used: "youden",
"Fmeasure","geo.mean". Default is "Fmeasure".}

\item{B:}{number of permutations.}
}
\value{
Four lists:
\itemize{
\item ACS: genome-wide c-scores.
\item ACSpvalue: p-values for genome-wide c-scores.
\item ADS: genome-wide d-scores.
\item ADSpvalue: p-values for genome-wide d-scores.
}
In addition, the four data matrices are written to a folder named "ACS_ADS_Global".
}
\description{
The \code{multi_ACS_ADS_global} is function to perform congruence analysis
for multiple study pairs, generating genome-wide c-scores, d-scores and their permuted p-values.
}
\examples{
\dontrun{
#mcmc.merge.list from the merge step (see the example in function 'merge')
dataset.names = c("hb","hs","ht","ha","hi","hl",
                  "mb","ms","mt","ma","mi","ml")
ACS_ADS_global = multi_ACS_ADS_global(mcmc.merge.list,dataset.names,
measure="Fmeasure",B=1000)
}
}
