% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ACS_ADS_pathway.R
\name{multi_ACS_ADS_pathway}
\alias{multi_ACS_ADS_pathway}
\title{Pathway level congruence analysis for multiple pairs}
\usage{
multi_ACS_ADS_pathway(
  mcmc.merge.list,
  dataset.names,
  select.pathway.list,
  measure = "Fmeasure",
  B = 100,
  parallel = F,
  n.cores = 4
)
}
\arguments{
\item{mcmc.merge.list:}{a list of merged MCMC output matrices.}

\item{dataset.names:}{a vector of dataset names.}

\item{measure:}{three types of scores to be used: "youden",
"Fmeasure","geo.mean". Default is "Fmeasure".}

\item{B:}{number of permutations.}

\item{parallel:}{whether to perform parallel ('mclapply' will be used if parallel=T) computing in permutation.}

\item{n.cores:}{if parallel=T, the number of cores to use (mc.cores parameter in 'mclapply' function).}
}
\value{
Four lists:
\itemize{
\item ACS.mat: pathway-specific c-scores.
\item ACSpvalue.mat: p-values for pathway-specific c-scores.
\item ADS.mat: pathway-specific d-scores.
\item ADSpvalue.mat: p-values for pathway-specific d-scores.
}
In addition, the four data matrices are written to a folder named "ACS_ADS_Pathway".
}
\description{
The \code{multi_ACS_ADS_pathway} is function to perform congruence analysis
for multiple pairs, generating pathway-specific c-scores, d-scores and their permuted p-value.
}
\examples{
\dontrun{
#mcmc.merge.list from the merge step (see the example in function 'merge')
#select.pathway.list from the pathSelect step (see the example in function 'pathSelect')
dataset.names = c("hb","hs","ht","ha","hi","hl",
                  "mb","ms","mt","ma","mi","ml")
ACS_ADS_pathway = multi_ACS_ADS_pathway(mcmc.merge.list,dataset.names,
                                        select.pathway.list,
                                        measure="Fmeasure",
                                        B=1000,parallel=F)
}
}
