% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicox.R
\name{multicox}
\alias{multicox}
\title{Title multicox calculate}
\usage{
multicox(form, data, names)
}
\arguments{
\item{form}{form a form for mulcox}

\item{data}{data.frame style contain variable which is factor}

\item{names}{mulcox variable}
}
\value{
multicox2 data.frame style contains results
}
\description{
Title multicox calculate
}
\examples{
form <- mul.form(c('gender','age'))(time="OS",status="EVENT",data=df)
          multicox(form,data=df,names=c("group",'gender','age'))

}
