% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOutput.R
\name{multiOutput}
\alias{multiOutput}
\title{Downstream visualization tools}
\usage{
multiOutput(
  mcmc.merge.list,
  dataset.names,
  select.pathway.list,
  ACS_ADS_pathway,
  output = c("clustPathway", "mdsModel", "clustModel", "genePM", "keggView",
    "reactomeView"),
  optK = NULL,
  sil_cut = 0.1,
  use_ADS = FALSE,
  hashtb = NULL,
  keywords_cut = 0.05,
  text.permutation = "all",
  comemberProb_cut = 0.7,
  ViewPairSelect = NULL,
  kegg.species = "hsa",
  KEGG.dataGisEntrezID = FALSE,
  KEGG.dataG2EntrezID = NULL,
  KEGG.pathID2name = NULL,
  reactome.species = "HSA",
  Reactome.dataG2TopologyGtype = NULL,
  Reactome.pathID2name = NULL
)
}
\arguments{
\item{mcmc.merge.list:}{a list of merged MCMC output matrices.}

\item{dataset.names:}{a vector of dataset names matched with the mcmc.merge.list.}

\item{select.pathway.list:}{a list of selected pathways (containing gene components) for clustering/visualizations.}

\item{ACS_ADS_pathway:}{a list of four data frames: pathway-specific c-scores, pathway-specific d-scores
and their permuted p-value (each row is a pathway and each column is a study).}

\item{output:}{choose from: "clustPathway" (pathway clustering),
"mdsModel"(within-pathway MDS plot on studies),
"clustModel" (within-pathway clustering heatmap),
"genePM" (within-pathway posterior DE heatmap),
"keggView" (KEGG pathway topology, default is human - hsa, for other species,
KEGG organism name and gene Entrez ID needs to be provided as 'KEGG.dataG2EntrezID'),
"reactomeView" (Reactome pathway topology, default is human - HSA, for other species, Reactome organism name needs
to be provided as "reactome.species").
Clustering analysis is not applicable when the number of studies is smaller than 3. "output" cannot be empty.}

\item{optK:}{Optimal number of clusters. For "clustPathway" output only.}

\item{sil_cut:}{silhouette index to control scatterness. Larger value indicates tigher cluster and
more scattered pathways.}

\item{use_ADS:}{whether use d-scores for clustering/visualizations. Default is FALSE.}

\item{hashtb:}{a flat noun-pathway table for text mining. Prepared tables from KEGG and Reactome pathway
descriptions for 5 species "hsa", "mmu", "rno", "cel" and "dme"are provided by \code{data(hashtb_hsa)},
\code{data(hashtb_mmu)}, \code{data(hashtb_rno)}, \code{data(hashtb_cel)} and \code{data(hashtb_dme)}.
Please refer to Zeng, Xiangrui, et al. "Comparative Pathway Integrator: a framework
of meta-analytic integration of multiple transcriptomic studies for consensual and differential pathway
analysis." Genes 11.6 (2020): 696.}

\item{keywords_cut:}{keywords above this cut will be shown in the text mining spreadsheet output.}

\item{text.permutation:}{select from "all" or "enriched". In text mining, "all" permutates pathways from
full pathway.list provided while "enriched" permutates from selected pathways. "all" is suitable for
cross-species comparision while "enriched" is recommended for within-species comparision.}

\item{comemberProb_cut:}{probability below this cut will be colored blue in comembership heatmaps.}

\item{ViewPairSelect:}{which two datasets to view in the KEGG/Reactome topology plot. All pairs will be
considered under default (may take a while).}

\item{kegg.species:}{KEGG species abbreviation. For "keggView" only. Default is "hsa".}

\item{KEGG.dataGisEntrezID:}{whether gene names in data are EntrezID. Default is FALSE.}

\item{KEGG.dataG2EntrezID:}{a data frame which maps gene names in mcmc.merge.list (first column) to
Entrez IDs (second column). If NULL & KEGG.dataGisEntrezID=F & kegg.species is one of "hsa", "mmu",
"rno", "cel" or "dme", gene symbols will be automatically mapped to EntrezID by Bioconductor packages
"org.Hs.eg.db", "org.Mm.eg.db", "org.Rn.eg.db", "org.Ce.eg.db" or org.Dm.eg.db". For "keggView" only.}

\item{KEGG.pathID2name:}{a list where each element is a KEGG pathway name with correponding pathway ID
as its name. ID will be retrieved from KEGGREST if this is NULL.}

\item{reactome.species:}{Reactome species abbreviation. For "reactomeView" only. Default is "HSA".}

\item{Reactome.dataG2TopologyGtype:}{a data frame which maps gene names in mcmc.merge.list (first
column) and select.pathway.list to gene name types in Reactome topology (second column). For "reactomeView" only.}

\item{Reactome.pathID2name:}{a list where each element is a Reactome pathway name with correponding
pathway ID as its name. ID will be retrieved from reactome.db if this is NULL.}
}
\value{
all figures and tables are stored in created folders in the current directory.
}
\description{
Downstream visualization tools: visualization outputs including overall
pathway clustering and output for each pathway
The \code{multiOutput} is function to visualize pathway level c-scores and d-scoresincluding
pathway clustering results with co-membership heatmaps,
within-pathway MDS plot (on studies),
within-pathway clustering heatmap (on studies),
within-pathway posterior DE heatmap,
KEGG and Reactome topology plots for each pathway.
}
\examples{
\dontrun{
#mcmc.merge.list from the merge step (see the example in function 'merge')
#select.pathway.list from the pathSelect step (see the example in function 'pathSelect')
#ACS_ADS_pathway from the multi_ACS_ADS_pathway step (see example in 'multi_ACS_ADS_pathway')
data(hashtb_hsa) #include hashtb & pathways for text mining
dataset.names = c("hb","hs","ht","ha","hi","hl",
                  "mb","ms","mt","ma","mi","ml")
#1. step1: select K by elbow plot from consensus clustering
ACSpvalue.mat = ACS_ADS_pathway$ACSpvalue.mat
results = ConsensusClusterPlus(d=t(-log10(ACSpvalue.mat)),maxK=10,reps=50,pItem=0.8,
pFeature=1,title="Consensus Clustering",clusterAlg="hc",innerLinkage="ward.D2",
finalLinkage="ward.D2",seed=12345,plot="png")

#2. step2: run multiOutput with pre-selected K=4
multiOutput(mcmc.merge.list,dataset.names,select.pathway.list,ACS_ADS_pathway,
           output=c("clustPathway","mdsModel","clustModel","genePM","keggView"),
           hashtb=hashtb,optK = 4,keywords_cut=0.2,comemberProb_cut=0.6)
}
}
