% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{An input data frame.}

\item{cl}{A true class value of your training data.}

\item{k_nn}{An integer representing the number of neighbors.}

\item{k_cv}{An integer representing the number of folds.}
}
\value{
A list containing a vector set of the predicted class for all the
  observatoin and a number representing the cross-validation
  misclassification error.
}
\description{
This function performs a k-nearest neighbors cross-validation on the given
  data.
}
\examples{
my_knn_cv(iris, iris$Species, 1, 5)

}
\keyword{prediction}
