% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear model function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A formula class object.}

\item{data}{Input data frame.}
}
\value{
A table which is similar to the summary() given by the lm() function.
1
}
\description{
This function fits a linear model in r and give a summary table.
}
\examples{
my_lm(mpg ~ hp + wt, data = mtcars)

}
\keyword{prediction}
