% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{A number of fold.}
}
\value{
A number representing the cross-validation error.
}
\description{
This function performs a random forest cross-validation on the given data.
}
\examples{
my_rf_cv(5)

}
\keyword{prediction}
