% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{A numeric set of data.}

\item{alternative}{A character string specifying the alternative.}

\item{mu}{A number indicating the null hypothesis value of the mean.}
}
\value{
A list containing the test statistic, degree of freedom, a string
  represent the type of t-test, and the p-value of the t-test.
}
\description{
This function performs a sample t-test based on the given input.
}
\examples{
my_t.test(sample, "less", 1)
my_t.test(sample, "greater", 1)
my_t.test(sample, "two.sided", 1)

}
\keyword{prediction}
