% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial experiment}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations of the experiment}

\item{n}{number of trails}

\item{p}{probability of a success}
}
\value{
a barplot showing the distribution of successes after n trials
}
\description{
Binomial experiment
}
\examples{
mybin(iter = 10000, n = 10, p = 0.7)
}
