% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{myboot}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{sample}

\item{fun}{mean/median}

\item{alpha}{0.05 default}

\item{...}{additional parameters}
}
\value{
A graph that plots the sample (x) and also displays the confidence intervalsSS
}
\description{
myboot
}
\examples{
myboot(iter=1000, x, fun="median", alpha = 0.01, cx=1.5)
}
