% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot2}
\alias{myboot2}
\title{myboot2 function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{- population for samples}

\item{x}{- sample of numbers}

\item{fun}{- mean median sd, etc}

\item{alpha}{- bootstrap interval}

\item{cx}{- character expansion}

\item{...}{}
}
\value{
hist or list
}
\description{
myboot2 function
}
