% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mychisim.R
\name{mychisim}
\alias{mychisim}
\title{Chi Squared Function Simulator}
\usage{
mychisim(n, sigma, mean, iter, ymax, ...)
}
\arguments{
\item{n}{the size of each sample}

\item{sigma}{standard deviation of the sample}

\item{mean}{mean for the sample}

\item{iter}{number of iterations}

\item{ymax}{maximum y value for sample}

\item{...}{additional labels for plot}
}
\value{
Chisq Plot of input sample, Chi-Squared Statistic w, summary for the
statistic, and the standard deviations of the statistic.
}
\description{
Takes in the size of sample, standard deviation of the sample,
mean of the sample, number of iterations of samples, and the maximum y value
for the sample and returns a histogram of the Chi Squared statistic for the
sample as well as a density plot containing a theoretical normal curve.
}
\examples{
mychisim(n=10,sigma=4,mean=10,iter=1000,ymax=0.1)

}
