% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{CLT Binomial Simulation}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{number of trails}

\item{iter}{number of samples to generate}

\item{p}{probability of a success}

\item{...}{extra parameters used by the histogram}
}
\value{
a histogram compared with a normal curve showing the CLT.
}
\description{
CLT Binomial Simulation
}
\examples{
mycltb(n = 10, iter = 10000, p = 0.7)
}
