% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{kuihkjhadkjsahdkjsahdkjashdkjsahdkjashdas
asdashkdhaskhdkashdkjashdkashdkjashdas
daskdbajsdkashdklahsdklsahldALDJSAòLJDAS}
\format{Data Frame}
\source{
read.csv(system.file("extdata/FileDse.csv", package = "MyPackage"),stringsAsFactors = FALSE)
}
\usage{
mydata
}
\description{
kuihkjhadkjsahdkjsahdkjashdkjsahdkjashdas
asdashkdhaskhdkashdkjashdkashdkjashdas
daskdbajsdkashdklahsdklsahldALDJSAòLJDAS
}
\keyword{datasets}
