% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydbinom.R
\name{mydbinom}
\alias{mydbinom}
\title{Density distribution function}
\usage{
mydbinom(a, b, c)
}
\arguments{
\item{a}{- vector quantities}

\item{b}{- number of observations.}

\item{c}{- probability of success on each trial}
}
\value{
density distributions
}
\description{
Density distribution function
}
\examples{
isaiahpackage::mydbinom(1:10,100,0.5)
}
