% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Max Likelihood Function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function to be applied}

\item{x}{vector containing a random sample}

\item{param}{extra parameter for lfun}

\item{...}{extra parameters used by plot}
}
\value{
plot showing where the maximum likelihood is
}
\description{
Max Likelihood Function
}
\examples{
\dontrun{
mymaxlik(x=c(4,6,7,6,5), param=seq(0,10,length=1000),lfun=logpoiss)
}
}
