% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlbeta.R
\name{mymlbeta}
\alias{mymlbeta}
\title{Maximum Likelihood Function for Beta Distribution}
\usage{
mymlbeta(x, alpha, beta, ...)
}
\arguments{
\item{x}{sample vector}

\item{alpha}{vector of alpha values}

\item{beta}{vector of beta values}

\item{...}{additional plot labels/settings}
}
\value{
Graph showing the maximum likelihood for Beta
}
\description{
A function that takes in a sample from the beta distribution along
with alpha and beta of the sample and produces a graph showing the maximum
likelihood function for the sample.
}
\examples{
sam<-rbeta(30,shape1=3,shape2=4)
nsam<-length(sam)
mymlbeta(x=sample(sam,nsam,replace=TRUE),alpha=seq(0.1,20,length=100),
beta=seq(0.1,20,length=100),lwd=2,labcex=1,col="steelblue")

}
