% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorn.R
\name{mymlnorm}
\alias{mymlnorm}
\title{Mymlnorm}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{sample vector}

\item{mu}{sequence of mu}

\item{sig}{sequence of sig}

\item{...}{}
}
\value{
graph of the likelyhood estimates for the population mean and standard deviation.
}
\description{
Mymlnorm
}
\examples{
mymlnorm(x,mu=10,sig=10)
}
