% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{My Plot}
\usage{
myplot(x)
}
\arguments{
\item{x}{The diameter}
}
\value{
Response ie estimate height
}
\description{
My Plot
}
\examples{
\dontrun{quad.lm <- lm(y~x, data.df);
myplot(x = 15, spruce.df)}

}
