% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomo.plot.R
\name{nomo.plot}
\alias{nomo.plot}
\title{Title nomo.plot}
\usage{
nomo.plot(
  variable,
  t1 = 1095,
  t2 = 1825,
  form,
  time = time,
  status = status,
  data = df
)
}
\arguments{
\item{variable}{character variable which is factor}

\item{t1}{3 year (day)}

\item{t2}{5 year (day)}

\item{data}{data.frame style contain variable which is factor}
}
\value{
plot
}
\description{
Title nomo.plot
}
\examples{
variable=names(df)[2:7]
form <- mul.form(variable)(time="OS",status="EVENT",data=df)
nomo.plot(variable,t1=1095,t2=1825,form,time="OS",status="EVENT",data=df)
}
