% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalizeFilename.R
\name{normalizeFilename}
\alias{normalizeFilename}
\title{Fix characters following the naming conventions [http://msdn.microsoft.com/en-us/library/windows/desktop/aa365247(v=vs.85).aspx]:}
\usage{
normalizeFilename(x, form = c("default", "8.3")[1], fix.encoding = TRUE,
  sub.sign = "")
}
\arguments{
\item{x}{the raw character string}

\item{form}{the format}

\item{fix.encoding}{whether encoding should be fixed}

\item{sub.sign}{character to use as substitution}
}
\value{
the fixed character string
}
\description{
Fix characters following the naming conventions [http://msdn.microsoft.com/en-us/library/windows/desktop/aa365247(v=vs.85).aspx]:
}

