% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Optimal Ticket Finding}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Variable for the amount of seats on the place}

\item{gamma}{Variable for the probability the plane is overbooked}

\item{p}{Variable for the probability that a passenger shows}
}
\value{
Creates two plots representing the Objective Vs n to find the optimal tickets
sold using the binomial distribution and normal approximation. Then creates a named
list containing the optimal tickets sold using both distributions along with N,
gamma, and p.
}
\description{
Optimal Ticket Finding
}
\examples{
ntickets(200, 0.02, 0.95)
ntickets(500, 0.01, 0.9)

}
