% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{Oldest Player Function}
\usage{
oldest_player(data, year)
}
\arguments{
\item{year}{Which is your year of interest? Range from 1950 to 2017.}
}
\description{
This function allows you to get the player earned the most points for a specific year.
}
\examples{
oldest_player()

}
\keyword{age}
