% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_plr.R
\name{oldest_plr}
\alias{oldest_plr}
\title{A Finding Player Function}
\usage{
oldest_plr(x)
}
\arguments{
\item{year}{Specified a year to return the oldest player at that time}
}
\description{
This function allows you to find the oldest NBA at the year specified.
If more than 1 players have the same age at that time, only one player will be shown (Alphabetical order)
}
\examples{
oldest_plr(1952)
}
\keyword{oldest}
