% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grDevices.R
\name{openPdfDevice}
\alias{openPdfDevice}
\title{Open PDF device
initialize to file name after title}
\usage{
openPdfDevice(title, height = NULL)
}
\arguments{
\item{title}{the title}

\item{height}{the plot height in inches}
}
\description{
Open PDF device
initialize to file name after title
}

