% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grDevices.R
\name{openPngDevice}
\alias{openPngDevice}
\title{Open PNG device
initialize to file name after title}
\usage{
openPngDevice(title, height = NULL, width = NULL, res = NULL)
}
\arguments{
\item{title}{the title}

\item{height}{the plot height in inches}

\item{width}{the plot width in inches}

\item{res}{the plot resolution in dots per inch}
}
\description{
Open PNG device
initialize to file name after title
}

