% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwaySelect.R
\name{pathSelect}
\alias{pathSelect}
\title{Select pathways of interest for pathway-level congruence analysis}
\usage{
pathSelect(
  mcmc.merge.list,
  pathway.list,
  pathwaysize.lower.cut = 10,
  pathwaysize.upper.cut = 200,
  overlapsize.cut = 10,
  med.de.cut = 5,
  min.de.cut = 0,
  qfisher.cut = 0.05,
  topPath.indStudy.num = NULL
)
}
\arguments{
\item{mcmc.merge.list:}{a list of merged MCMC output matrices.}

\item{pathway.list:}{list of pathway database. This can be any pathway list from external
sources or selected from the contained pathway data in CAMO package.}

\item{pathwaysize.lower.cut:}{pathway size lower bound cutoff;}

\item{pathwaysize.upper.cut:}{pathway size upper bound cutoff;}

\item{overlapsize.cut:}{the lower bound cutoff of overlap size between genes
from input data and genes from a pathway.}

\item{med.de.cut:}{the lower bound cutoff of minimum number of DE genes in a
pathway.}

\item{qfisher.cut:}{fisher q-value cutoff from the meta enrichment
analysis.}

\item{topPath.indStudy.num:}{if not NULL, only the union of top pathways is
considered.}
}
\value{
a vector of selected pathway names.
}
\description{
The \code{pathSelect} is function to select pathways of interest for
pathway-level congruence analysis by meta pathway enrichement analysis.
}
\examples{
\dontrun{
#mcmc.merge.list from the merge step (see the example in function 'merge')
data(human.pathway.list) ## include pathway.list
select.pathway = pathSelect(mcmc.merge.list,pathway.list,
                            pathwaysize.lower.cut = 5,
                            pathwaysize.upper.cut=200,
                            overlapsize.cut = 5, med.de.cut =3,
                            qfisher.cut = 0.05)
select.pathway.list = pathway.list[select.pathway]
}
}
