% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pays}
\alias{pays}
\title{Jeu de données sur les pays du monde}
\format{
Un objet data.frame avec 241 lignes et 6 colonnes
}
\usage{
pays
}
\description{
Jeu de données sur les pays du monde
}
\keyword{datasets}
