% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{methods}
\name{penjelas}
\alias{penjelas}
\title{Get Predictor Variables}
\usage{
penjelas(model)
}
\arguments{
\item{model}{An object of class \code{RLB}.}
}
\value{
A numeric matrix representing the predictor variables X.
}
\description{
Extracts the predictor variables matrix from an RLB object.
}
