% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{methods}
\name{penjelas}
\alias{penjelas}
\title{Ekstraktor untuk Matriks Penjelas dari Objek RLB}
\usage{
penjelas(model)
}
\arguments{
\item{model}{Objek dari kelas \code{RLB}.}
}
\value{
\code{matrix} Matriks variabel penjelas.
}
\description{
Mengembalikan matriks variabel penjelas (X) dari objek kelas RLB.
}
