% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{penjelas}
\alias{penjelas}
\title{Aksesor untuk Matriks Penjelas (X)}
\usage{
penjelas(model)
}
\arguments{
\item{model}{Objek kelas RLB.}
}
\value{
Mengembalikan slot X.
}
\description{
Mengambil slot X dari objek RLB.
}
