% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise.R
\name{piecewise}
\alias{piecewise}
\title{Piecewise linear function}
\usage{
piecewise(x, xk, coef)
}
\arguments{
\item{x}{X variable}

\item{xk}{X coordinate of intersection}

\item{coef}{Coefficients of the function as a vector}
}
\value{
A piecewise linear function with respect to X
}
\description{
Piecewise linear function
}
\examples{
piecewise(5, 8, c(0.5, 0.2, -0.3))
}
