% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{methods}
\name{plot,RLB,missing-method}
\alias{plot,RLB,missing-method}
\title{Plot Method for RLB Objects}
\usage{
\S4method{plot}{RLB,missing}(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{RLB}.}

\item{y}{Missing, not used in this method.}

\item{...}{Additional arguments passed to the plot function.}
}
\description{
Generates scatter plots of each predictor variable against the response variable,
with the simple linear regression line overlaid.
}
