% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot,RLB,missing-method}
\alias{plot,RLB,missing-method}
\title{Metode Plot untuk Objek RLB}
\usage{
\S4method{plot}{RLB,missing}(x, y, ...)
}
\arguments{
\item{x}{Objek kelas RLB.}

\item{y}{Parameter `y` diabaikan untuk method ini.}

\item{...}{Argumen tambahan untuk plot.}
}
\description{
Membuat plot sebar untuk setiap variabel X terhadap y.
}
