% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot,RLB,missing-method}
\alias{plot,RLB,missing-method}
\title{Plot untuk Objek RLB}
\usage{
\S4method{plot}{RLB,missing}(x, y, ...)
}
\arguments{
\item{x}{Objek dari kelas \code{RLB}.}

\item{y}{Argumen ini tidak digunakan, disertakan untuk konsistensi dengan generic plot.}

\item{...}{Argumen tambahan untuk fungsi plot.}
}
\value{
Tidak mengembalikan nilai, menghasilkan plot.
}
\description{
Membuat scatter plot untuk setiap variabel penjelas terhadap variabel respons
dari objek kelas RLB, beserta garis regresi sederhana.
}
