% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_perplexity.R
\name{plot_perplexity}
\alias{plot_perplexity}
\title{plot_perplexity}
\usage{
plot_perplexity(data.frame, value.write, xintercept.write, logx)
}
\arguments{
\item{data.frame}{Dataframe. Perplexity values returned by fit.topics.perplexity(), fit.alpha.perplexity() and fit.beta.perplexity()}

\item{value.write}{Character. What perplexity is calculated for; either 'K', 'Alpha' or 'Beta'}

\item{logx}{Logical. Whether to plot the x axis logarithmically (often useful for Beta and Alpha plots)}

\item{x.intercept.write}{Numeric. Optional parameter: value of x-intercept to plot a vertical dotted red line showing chosen value}
}
\value{
ggplot2 plot showing the values of perplexity (both each fold and the average)
}
\description{
Plots the perplexity values for K, alpha and beta, showing each fold as a point and the average as a line
}
