% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_of_yr.R
\name{pos_of_yr}
\alias{pos_of_yr}
\title{A Tom Van Function}
\usage{
pos_of_yr(year)
}
\arguments{
\item{year}{Which year are you interested in?}
}
\description{
This function retrives the year Tom Van played in the year specified
}
\examples{
pos_of_yr(1966)
}
\keyword{pos}
