% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{rand}
\alias{rand}
\title{Random numbers}
\usage{
rand(n, mean = 0, sd = 1, style = c("vector", "data.frame"),
  ncol = NULL)
}
\arguments{
\item{n}{number of observations.}

\item{mean}{distribution mean}

\item{sd}{standard deviation of the mean}

\item{style}{the style output ("vector", "data.frame")}

\item{ncol}{the number of columns if style == "data.frame"}
}
\value{
The function \code{rand} returns a a vector or a data.frame with
multiple random numbers.
components:

\code{vector}: A numeric vector with n observations

\code{data.frame}: A data.frame with n rows and ncol columns. Each
colulm corresponds to one vector with n observations.
}
\description{
Generate random sequences of numbers using normal distribution
}
\details{
This functions generate random numbers based on the nomral distribution
through the function \code{rnorm}. It uses the 'mean' and standard deviation of the
mean to generate 'n' random numbers. The user can choose to generate multiple
vectors by setting style == "data.frame".
}
\examples{
rand(n = 10)
rand(n = 100, mean = 3, sd = 1)
# style == "data.frame"
rand(n = 10, style = "data.frame", ncol = 1)
rand(n = 100, mean = 3, sd = 1, style = "data.frame", ncol = 5)
}
\references{
Your refrence goes here.
}
\seealso{
\code{\link[stats]{rnorm}}
}
\author{
Your Full Name
}
