% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{read.pData}
\alias{read.pData}
\title{Read p-value data.}
\usage{
read.pData(p.file, sep = ",", quote = "\\"", header = T)
}
\arguments{
\item{p.file:}{the file name of the group data to be read.}

\item{sep:}{the field separator character. Values on each line of
the file are separated by this character. If \code{sep = ""}, the
separator is 'white space', that is one or more spaces, tabs,
newlines or carriage returns. The default is \code{,}. See
\link{read.csv}.}

\item{quote:}{the set of quoting characters. To disable quoting
altogether, use \code{quote = ""}.  See \link{scan} for the
behaviour on quotes embedded in quotes. Quoting is only
considered for columns read as character, which is all of them
unless colClasses is specified.  The default is \code{\"}. See
\link{read.csv}}

\item{header:}{a logical value indicating whether the file contains
the names of the variables as its first line. If missing, the
value is determined from the file format: \code{header} is set to
\code{TRUE} if and only if the first row contains one fewer field
than the number of columns. The default is \code{TRUE}. See
\link{read.csv}.}
}
\value{
the group label vector.
}
\description{
Read p-value data
The \code{read.pData} is function to read p-value data
}
\examples{
\dontrun{
data(hb)
summaryDE <- indDE(data=data,group=group,data.type="microarray",
                  case.label="2", ctrl.label="1")
pData <- summaryDE[,c(3,1)]
write.csv(pData,file="pData.csv")
pData <- read.pData(p.file="pData.csv")
}
}
