% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readData.R
\name{readDataCsv}
\alias{readDataCsv}
\title{Read CSV data, enforce string representation}
\usage{
readDataCsv(csvFilePath)
}
\arguments{
\item{csvFilePath}{file path to CSV}
}
\value{
data frame
}
\description{
Read CSV data, enforce string representation
}

