% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_videos.R
\name{region_to_id}
\alias{region_to_id}
\title{Converts region to YouTube id}
\usage{
region_to_id(key, region)
}
\arguments{
\item{key}{Your YouTube API key}

\item{region}{String containing region of intrest.}
}
\value{
A string of a single region id
}
\description{
Converts region to YouTube id
}
