% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regnomo.plot.R
\name{regnomo.plot}
\alias{regnomo.plot}
\title{Title regnomo.plot}
\usage{
regnomo.plot(variable, time = "OS", status = "EVENT", form, t1, t2, data = df)
}
\arguments{
\item{variable}{character variable which is factor}

\item{time}{survival time}

\item{status}{survival status}

\item{data}{data.frame style contain variable which is factor}
}
\value{

}
\description{
Title regnomo.plot
}
\examples{
variable <- names(df)[2:5]
form <- mul.form(variable)(time="OS",status="EVENT",data=df)
regnomo.plot(variable,t1=1095,t2=1825,form=form,time="OS",status="EVENT",data=df)
}
