% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacePunctuation.R
\name{replacePunctuation}
\alias{replacePunctuation}
\title{replacePunctuation}
\usage{
replacePunctuation(x, ...)
}
\arguments{
\item{corpus}{Corpus from the tm package. Created by using tm::Corpus(VectorSource([text data]))}
}
\value{
Cleaned text
}
\description{
Replaces punctuation with a space. Should be used with the tm::tm_map command.  Can lead to better results than removePunctuation
}
