% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceSingleWords.R
\name{replaceSingleWords}
\alias{replaceSingleWords}
\title{replaceSingleWords}
\usage{
replaceSingleWords(x, ...)
}
\arguments{
\item{corpus}{Corpus from the tm package. Created by using tm::Corpus(VectorSource([text data]))}
}
\value{
Cleaned text
}
\description{
Replaces single words with a space. Should be used with the tm::tm_map command. Advisable to cleanup text after other tm_map functions have been used
}
