% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{residuals,RLB-method}
\alias{residuals,RLB-method}
\title{Metode Residuals untuk Objek RLB}
\usage{
\S4method{residuals}{RLB}(object)
}
\arguments{
\item{object}{Objek kelas RLB.}
}
\value{
Residual dari model.
}
\description{
Menghitung residual dari model RLB.
}
