% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{methods}
\name{residuals,RLB-method}
\alias{residuals,RLB-method}
\title{Residuals Method for RLB Objects}
\usage{
\S4method{residuals}{RLB}(object, ...)
}
\arguments{
\item{object}{An object of class \code{RLB}.}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric vector representing the residuals (y - y_hat).
}
\description{
Calculates the residuals for an RLB model.
}
