% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{residuals,RLB-method}
\alias{residuals,RLB-method}
\title{Residuals untuk Objek RLB}
\usage{
\S4method{residuals}{RLB}(object)
}
\arguments{
\item{object}{Objek dari kelas \code{RLB}.}

\item{...}{Argumen tambahan (tidak digunakan).}
}
\value{
\code{numeric} Vektor residual.
}
\description{
Menghitung residual dari model regresi linier berganda yang disimpan
dalam objek kelas RLB.
}
