% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{methods}
\name{respons}
\alias{respons}
\title{Get Response Variable}
\usage{
respons(model)
}
\arguments{
\item{model}{An object of class \code{RLB}.}
}
\value{
A numeric vector representing the response variable y.
}
\description{
Extracts the response variable from an RLB object.
}
