% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{methods}
\name{respons}
\alias{respons}
\title{Ekstraktor untuk Vektor Respons dari Objek RLB}
\usage{
respons(model)
}
\arguments{
\item{model}{Objek dari kelas \code{RLB}.}
}
\value{
\code{numeric} Vektor respons.
}
\description{
Mengembalikan vektor respons (y) dari objek kelas RLB.
}
