% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{respons}
\alias{respons}
\title{Aksesor untuk Variabel Respons (y)}
\usage{
respons(model)
}
\arguments{
\item{model}{Objek kelas RLB.}
}
\value{
Mengembalikan slot y.
}
\description{
Mengambil slot y dari objek RLB.
}
