% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{class}
\name{RLB-class}
\alias{RLB-class}
\title{RLB Class}
\description{
A class to represent a Multiple Linear Regression model.
}
\section{Slots}{

\describe{
\item{\code{y}}{A numeric vector representing the response variable.}

\item{\code{X}}{A numeric matrix representing the predictor variables.}

\item{\code{koef}}{A numeric vector representing the estimated coefficients.}
}}

