% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RLB}
\alias{RLB}
\title{Constructor for RLB Class}
\usage{
RLB(y, ...)
}
\arguments{
\item{y}{A numeric vector for the response variable.}

\item{...}{One or more numeric vectors or a matrix for the predictor variables.
These will be column-binded to form the predictor matrix X.}
}
\value{
An object of class \code{RLB}.
}
\description{
Creates an object of class RLB representing a Multiple Linear Regression model.
}
\examples{
\dontrun{
set.seed(123)
n <- 100
x1_data <- rnorm(n)
x2_data <- rnorm(n)
y_data <- 5 + 2*x1_data - 3*x2_data + rnorm(n)
model <- RLB(y_data, x1_data, x2_data)
print(model)
}
}
