% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{class}
\name{RLB_stepwise-class}
\alias{RLB_stepwise-class}
\title{RLB_stepwise Class}
\description{
A class that inherits from RLB and implements stepwise variable selection
for multiple linear regression.
}
\section{Slots}{

\describe{
\item{\code{y}}{A numeric vector representing the response variable.}

\item{\code{X}}{A numeric matrix representing the predictor variables selected by stepwise regression.}

\item{\code{koef}}{A numeric vector representing the estimated coefficients for the selected model.}
}}

