% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RLB_stepwise}
\alias{RLB_stepwise}
\title{Constructor for RLB_stepwise Class}
\usage{
RLB_stepwise(y, ...)
}
\arguments{
\item{y}{A numeric vector for the response variable.}

\item{...}{One or more numeric vectors or a matrix for the potential predictor variables.
These will be column-binded to form the initial predictor matrix.}
}
\value{
An object of class \code{RLB_stepwise}.
}
\description{
Creates an object of class RLB_stepwise, performing stepwise
variable selection before fitting the Multiple Linear Regression model.
}
\examples{
\dontrun{
set.seed(123)
n <- 100
x1_data <- rnorm(n)
x2_data <- rnorm(n)
x3_data <- rnorm(n)
y_data <- 5 + 2*x1_data + 0*x2_data + 3*x3_data + rnorm(n)

model_sw <- RLB_stepwise(y_data, x1_data, x2_data, x3_data)
summary(model_sw)
plot(model_sw)
}
}
