% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPackage.R
\name{RPackage}
\alias{RPackage}
\title{Use the xkcd API}
\usage{
RPackage(number)
}
\description{
The \code{RPackage} function calls the xkcd JSON API and returns
metadata about the comic in the form of a list
}
