% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_wrapper.R
\name{RSADecrypt}
\alias{RSADecrypt}
\title{RSADecrypt}
\usage{
RSADecrypt(ciphertext, d, n)
}
\arguments{
\item{ciphertext}{\itemize{
\item the string of characters to be decrypted
}}

\item{d}{\itemize{
\item the private key decryption component
}}

\item{n}{\itemize{
\item the key product component
}}
}
\value{
plaintext - the decrypted ciphertext as a character vector
}
\description{
RSADecrypt
}
\examples{
RSADecrypt("ͶҖĥĥă ȓărĥ®ԋ", 577, 1517)
}
