% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_wrapper.R
\name{RSAEncrypt}
\alias{RSAEncrypt}
\title{RSAEncrypt}
\usage{
RSAEncrypt(plaintext, e, n)
}
\arguments{
\item{plaintext}{\itemize{
\item the vector of characters to be encrypted
}}

\item{e}{\itemize{
\item the public key encryption component
}}

\item{n}{\itemize{
\item the key product component
}}
}
\value{
ciphertext - the encrypted plaintext
}
\description{
RSAEncrypt
}
\examples{
RSAEncrypt(plaintext, 2, 14)
}
