% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_wrapper.R
\name{RSAGenerateKey}
\alias{RSAGenerateKey}
\title{RSAGenerateKey}
\usage{
RSAGenerateKey(p = 2, q = 7)
}
\arguments{
\item{p}{\itemize{
\item optional first prime number
}}

\item{q}{\itemize{
\item optional second prime number
}}
}
\value{
keys = (n, e, d) in which n is the key product component, e is the public key encryption component, and d is the private key decryption component
}
\description{
RSAGenerateKey
}
