% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_wrapper.R
\name{RSAPrimeGenerate}
\alias{RSAPrimeGenerate}
\title{RSAPrimeGenerate}
\usage{
RSAPrimeGenerate(n = 6)
}
\arguments{
\item{n}{\itemize{
\item The number of bits for the prime number to have. This should be represented as a positive scalar integer value
}}
}
\value{
list containing two unique prime numbers of n bit size
}
\description{
RSAPrimeGenerate
}
\examples{
 RSAPrimeGenerate(6)
}
