% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/runMain.R
\name{runMain}
\alias{runMain}
\title{Run read plate layout}
\usage{
runMain(dir = NULL, ...)
}
\arguments{
\item{dir}{directory}

\item{...}{further arguments
see sampleConfig() for a config}
}
\description{
Run read plate layout
}

