% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/config.R
\name{sampleConfig}
\alias{sampleConfig}
\title{Load a sample configuration
Avoiding overwriting existing variables}
\usage{
sampleConfig()
}
\description{
Load a sample configuration
Avoiding overwriting existing variables
}

