\name{SC19057-package}
\alias{SC19057-package}
\alias{SC19057}
\docType{package}
\title{
  This package provides several functions for benchmarking R.
}
\description{
  R versions of knnde and indTest function are provided in this package (\code{knnde} and \code{indTest})
}
\details{
  To use function "knnde", make sure you've installed R_package "FNN", and to use "indTest", make sure you've installed R_package "kedd"
}
\author{
Liu Haijiao <hjliuu@mail.ustc.edu.cn>
}
\keyword{ package }
