% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreDensityForStudent.R
\name{ScoreDensityForStudent}
\alias{ScoreDensityForStudent}
\title{Analysis scores}
\usage{
ScoreDensityForStudent()
}
\value{

}
\description{
Takes in a dataframe and the list of students, calculates  classification
}
