% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_pop.R
\name{sd_pop}
\alias{sd_pop}
\title{sd_pop}
\usage{
sd_pop(x)
}
\arguments{
\item{x}{a vector of numeric values}
}
\value{
the standard deviation of the input is returned.
}
\description{
Calculates the standard deviation of a vector using n as the denominator (suitable for populations) rather than n-1 (suitable for samples).
}
