% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MausmiS.R
\name{square}
\alias{square}
\title{Square a given number}
\usage{
square(x)
}
\arguments{
\item{x}{A numeric value to be squared.}
}
\value{
The square of the number
}
\description{
Takes any numeric value and returns the square of this number.
}
\examples{
squared_number <- square(c(5))
squared_number <- square(c(23))
}
