% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squareroot.R
\name{squareroot}
\alias{squareroot}
\title{Square root a number}
\usage{
squareroot(x)
}
\arguments{
\item{x}{A numeric value to be square rooted}
}
\value{
The square root of the input
}
\description{
Takes in any numeric value and square root it
}
