% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemText.R
\name{stemText}
\alias{stemText}
\title{stemText}
\usage{
stemText(x, ...)
}
\arguments{
\item{corpus}{Corpus from the tm package. Created by using tm::Corpus(VectorSource([text data]))}
}
\value{
Cleaned text
}
\description{
Basic stemming of text; removes 'ed', 'es' and 'ing' endings from words. Can lead to better results than stemDocument
}
