% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stepwise_selection}
\alias{stepwise_selection}
\title{Fungsi Internal untuk Seleksi Variabel Stepwise}
\usage{
stepwise_selection(X, y)
}
\arguments{
\item{X}{\code{matrix} Matriks dari semua variabel penjelas kandidat.}

\item{y}{\code{numeric} Vektor respons.}
}
\value{
\code{matrix} Matriks variabel penjelas yang terpilih. Jika tidak ada
variabel yang terpilih, mengembalikan matriks dengan 0 kolom.
}
\description{
Melakukan seleksi variabel maju-mundur (both) berdasarkan kriteria AIC.
}
\keyword{internal}
