% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudentDistibution.R
\name{StudentsDistribution}
\alias{StudentsDistribution}
\title{StudentsDistribution}
\usage{
StudentsDistribution()
}
\value{
map of the students' nationality for this class
}
\description{
Enumerates students by nationality
}
