% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary,RLB-method}
\alias{summary,RLB-method}
\title{Metode Summary untuk Objek RLB}
\usage{
\S4method{summary}{RLB}(object)
}
\arguments{
\item{object}{Objek kelas RLB.}
}
\value{
Mencetak summary dan tidak mengembalikan nilai eksplisit.
}
\description{
Menampilkan ringkasan statistik model RLB.
}
