% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumplot.R
\name{sumplot}
\alias{sumplot}
\title{Title}
\usage{
sumplot(tbhdat, Scientificname)
}
\arguments{
\item{tbhdat}{input data}

\item{Scientificname}{chr string of species to plot}
}
\value{
A plot
}
\description{
Title
}
\examples{
tbhdat <- read.csv('data/example_len_dat.csv', stringsAsFactors = F)
sumplot(tbhdat, 'Lagodon rhomboides')
}
