% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_titles_sentiment_analysis.R
\name{titles_sentiment_contribution_bing}
\alias{titles_sentiment_contribution_bing}
\title{Generates a plot of top 10 words that contribute to positive and negative sentiment in the top
200 YouTube titles using the Bing lexicon}
\usage{
titles_sentiment_contribution_bing(api_key)
}
\arguments{
\item{API_key}{A user's API key}
}
\value{
A plot of top 10 words that contribute to positive and negative sentiment in the top 200 YouTube titles
}
\description{
Generates a plot of top 10 words that contribute to positive and negative sentiment in the top
200 YouTube titles using the Bing lexicon
}
