% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_titles_sentiment_analysis.R
\name{titles_sentiment_type_nrc}
\alias{titles_sentiment_type_nrc}
\title{Generates a plot of the number of words that contribute to different sentiments in  the top
200 YouTube titles using the NRC lexicon}
\usage{
titles_sentiment_type_nrc(api_key)
}
\arguments{
\item{api_key}{A user's API key}
}
\value{
A plot of the number of words that contribute to different sentiments for the top 200 YouTube titles
}
\description{
Generates a plot of the number of words that contribute to different sentiments in  the top
200 YouTube titles using the NRC lexicon
}
