% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic.coherence.R
\name{topic.coherence}
\alias{topic.coherence}
\title{topic.coherence}
\usage{
topic.coherence(ldaOut, dtm, n.terms)
}
\arguments{
\item{ldaOut}{fitted topic model}

\item{dtm}{document term matrix for the fitted topic model}

\item{n.terms}{The number of terms in each topic to calculate topic coherence on}
}
\value{
vector of topic coherence scores (same length as the number of topics in the model)
}
\description{
Calculates topic coherence for each topic in a topic model. Uses the implementation put forward by Mimno et al. (2011) - Note, this directly uses the code in the package 'SpeedyReader'
}
