% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oopCluster.R
\name{trace.mmcluster}
\alias{trace.mmcluster}
\title{building trace}
\usage{
\method{trace}{mmcluster}(m, numSteps = 1e+05, stopTime = Inf)
}
\arguments{
\item{m}{model queueing system}

\item{numSteps}{amount arrival customers}

\item{stopTime}{modeling time}
}
\value{
numeric list of the accumulated parameters
}
\description{
building trace
}
