% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarcox.R
\name{univarcox}
\alias{univarcox}
\title{Title univarcox}
\usage{
univarcox(x)
}
\arguments{
\item{x}{variable,character}

\item{time}{survival time}

\item{status}{survival status}

\item{data}{data.frame style contain variable which is factor}
}
\value{
univarcox data.frame style contains results
}
\description{
Title univarcox
}
\examples{
univarcox("age")(time="OS",status="EVENT",data=df)
univar<-lapply(variable_names, function(x)univarcox(x)(time="OS",status="EVENT",data=df))
univartable<-do.call(rbind,lapply(univar,data.frame))
}
