% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{data}
\name{useful_data}
\alias{useful_data}
\title{Edgar Anderson's Iris Data}
\format{
Data frame
\describe{
\item{sepal_width}{Width of the sepal (numeric)}
\item{sepal_length}{Length of the sepal (numeric)}
\item{petal_width}{Width of the petal (numeric)}
\item{petal_length}{Length of the petal (numeric)}
\item{species}{Species (factor with levels setosa, versicolor, virginica)}
}
}
\usage{
useful_data
}
\description{
With cleaner names...
}
\keyword{datasets}
