% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_titles_word_cloud.R
\name{valid_api_call}
\alias{valid_api_call}
\title{Determines if user is providing a valid API key}
\usage{
valid_api_call(api_key)
}
\arguments{
\item{api_key}{The Youtube API key}
}
\value{
A status code of the API call
}
\description{
Determines if user is providing a valid API key
}
