% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_videos.R
\name{validate_category}
\alias{validate_category}
\title{Validates category inputted by user}
\usage{
validate_category(key, region, category)
}
\arguments{
\item{key}{Your YouTube API key}

\item{region}{YouTube content region}

\item{category}{String or numeric value inputted by user}
}
\value{
String indicating if region given is valid or invalid
}
\description{
Validates category inputted by user
}
