% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_videos.R
\name{validate_region}
\alias{validate_region}
\title{Validates region inputted by user}
\usage{
validate_region(key, region)
}
\arguments{
\item{key}{Your YouTube API key}

\item{region}{String or numeric value inputted by user}
}
\value{
String indicating if region given is valid or invalid
}
\description{
Validates region inputted by user
}
