% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.select.R
\name{var.select}
\alias{var.select}
\title{Title variable select from univarcox}
\usage{
var.select(data, p)
}
\arguments{
\item{data}{univarcox data}

\item{p}{pvalue}
}
\value{
variable
}
\description{
Title variable select from univarcox
}
\examples{
tmp <- univarcox("age")(time="OS",status="EVENT",data=df)
          var.select(tmp,0.05)
}
