% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedfinaldegree.R, R/weightedaverage.R
\name{weightedaverage}
\alias{weightedaverage}
\title{Adjusted Final Degree Classification}
\usage{
weightedaverage(DfDSE, Students)

weightedaverage(DfDSE, Students)
}
\arguments{
\item{DfDSE}{it's the datframe that contains students data}

\item{score}{it's one possible mark between 18 and 30 that you can take in your missing exams}
}
\value{
adjusted final degree classification

result
}
\description{
Takes in a data frame and a possible score to all the exams that you still need to do and calculates your "adjusted" final degree
}
