% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpower.R
\name{xpower}
\alias{xpower}
\title{raise a number to power}
\usage{
xpower(x, y)
}
\arguments{
\item{x}{A numeric value to be raised to power}

\item{y}{numeric value, the power}
}
\value{
x power y
}
\description{
Takes x and raise it to the power y
}
