% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_videos.R
\name{category_to_id}
\alias{category_to_id}
\title{Converts content category to YouTube id}
\usage{
category_to_id(key, category, region = "US")
}
\arguments{
\item{key}{Your YouTube API key}

\item{category}{Full name of YouTube content category as string. Run get_category_list() to see list of YouTube content categories.}

\item{region}{Region. Default is "US". Run get_region_list() to see list of YouTube cntent regions.}
}
\value{
A string of a single category id
}
\description{
Converts content category to YouTube id
}
