% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_scatter.R
\name{cor_scatter}
\alias{cor_scatter}
\title{cor_scatter}
\usage{
cor_scatter(x, y = NULL, lags)
}
\arguments{
\item{x}{a vector of numeric values}

\item{y}{a vector of numeric values}

\item{lags}{a list containing the values to be lagged}
}
\value{
a lattice of scatter plots.
}
\description{
Plots separate scatter plots in a lattice for several lag values
}
