% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorPlots_fun.R
\name{error_bars}
\alias{error_bars}
\title{error_bars}
\usage{
error_bars(df, s = NULL, draw = "both")
}
\arguments{
\item{df}{name of the dataframe that you want to analyse. This MUST be of the format data.frame(x,y), otherwise it will NOT work.}

\item{s}{sequence of values (generated by seq(x,y,z)) - optional input: if not provided it will be calculated automatically.}

\item{draw}{optional input to draw disperseion on the error plot; can take values 'standard error', 'standard deviation' or 'both'. By default it takes 'both'.}
}
\value{
an 'error_bars.out' object which contains (i) a dataframe of useful summary variables (ii) a dot plot showing the error bars, and (iii) a copy of the original dataframe with the bin_x variable appended as a third column.
}
\description{
analysis of error bars for a variable
}
