% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot FARS Accidents on a State Map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer representing the state number.}

\item{year}{An integer or string representing the year.}
}
\value{
A map with accident locations plotted, or NULL if no data is available.
}
\description{
Plots the locations of accidents for a given state and year on a map.
}
\examples{
\dontrun{
fars_map_state(1, 2013)
}
}
