% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions5.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot a map of accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state number (character) for which the user wishes to plot a map of accidents}

\item{year}{The year (numeric) for which the user wishes to plot a map of accidents}
}
\value{
Returns a plot, which is a map of the specified state (i.e. \code{state.num}) showing the 
 geographic locations of accidents during the specified \code{year}
}
\description{
This function plots a map of accidents' locations which occurred within a 
specified year in a given state.
}
\examples{
\dontrun{fars_map_state('34',2013)}

}
