% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS Data File}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the file to read.}
}
\value{
A tibble containing the data from the file.
}
\description{
Reads a CSV file containing FARS data and returns it as a tibble.
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
}
}
