% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads a dataframe}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character. This is the name of the file which the user wishes to 
read as an R dataframe}
}
\value{
This function returns an R tibble for the provided file
}
\description{
This function takes a file name as an argument and checks if it 
exists before reading it as an R dataframe
}
\note{
This function may result in a warning message because the \code{tbl_df()} 
 function in dplyr 1.0.0 was deprecated.
}
\examples{
\dontrun{fars_read(my_data) #define my_data as the name of the file you wish to read as
 an R dataframe}
\dontrun{fars_read("accident_2013.csv.bz2")}

}
