% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions3.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Obtaining monthly information of accidents}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{This is a numeric vector of the years for which the user is interested
in deriving information from}
}
\value{
This function return a list of tibbles (with two columns, the \code{year}
 and \code{MONTH}) for each of the provided years
}
\description{
This function takes a vector of years as an argument and provides a list, each
element of the list is a data frame (corresponding to one element in the vector) 
with two columns, the \code{year} and \code{MONTH} that correspond to each accident
}
\note{
This function intends to return a list of tipple (with two columns, 
 the \code{year} and \code{MONTH}) for the provided years. The function returns 
 a list of \code{NULL} values with a warning message.
}
\examples{
fars_read_years(c(2013,2014))

}
