% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS Data by Month and Year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric or character vector of years.}
}
\value{
A tibble with months as rows and years as columns, showing the number of accidents.
}
\description{
Summarizes the number of accidents per month for each year.
}
\examples{
\dontrun{
fars_summarize_years(c(2013, 2014, 2015))
}
}
