% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions4.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes number of accidents}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{This is a numeric vector of the years for which the user is interested
in deriving information from}
}
\value{
Returns a tibble, which is a summary table of the count of accidents (e.g. fatal injuries 
 suffered in motor vehicle traffic crashes) per month for the years provided.
}
\description{
This function takes a vector of specific years as an argument and returns a 
summary of the count (e.g. fatal injuries in motor vehicle traffic crashes) per 
month within the years provided
}
\note{
This function's code runs the \code{\link{fars_read_years}} function, 
 so any errors or warnings in \code{\link{fars_read_years}} will affect the 
 performance of this function
}
\examples{
fars_summarize_years(c(2013,2014))

}
