% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi.R
\docType{data}
\name{hi}
\alias{hi}
\title{Gene expression data of human infection}
\format{
a matrix where rows are genes and columns are samples
}
\usage{
hi
}
\description{
The data has been preprocessed.
}
\keyword{datasets}
