% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_summary.R
\name{lib_summary}
\alias{lib_summary}
\title{R Library Summary}
\usage{
lib_summary()
}
\value{
A data.frame containing the count of packages in each of the user's
libraries
}
\description{
This is now official hahaha.
}
\examples{
lib_summary()
}
