% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create FARS Filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or string representing the year.}
}
\value{
A character string with the filename in the format "accident_\if{html}{\out{<year>}}.csv.bz2".
}
\description{
Generates a standard FARS filename for a given year.
}
\examples{
make_filename(2015)
}
