% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions2.R
\name{make_filename}
\alias{make_filename}
\title{Specify the file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{This is the year for which the user wishes to obtain data from}
}
\value{

}
\description{
This function takes as an argument the year for which the user wishes to 
acquire data and accordingly creates the associated file name
}
\examples{
make_filename(2013)
make_filename(2014)

}
