% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{Bootstrap Function for Re-sampling}
\usage{
myboot(x, iter, fun, alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{x}{data used to create samples}

\item{iter}{number of trials to be performed}

\item{fun}{function to apply to sample matrix}

\item{alpha}{1-alpha is the percent of the confidence interval}

\item{cx}{the amount by which text and symbols on graph should be magnified}

\item{...}{additional labels and settings for plot}
}
\value{
Histogram of the sample statistic function, confidence interval for
the data, function used for the statistic, and the original sample vector
}
\description{
Create bootstrapped samples of any distribution by supplying the sample vector, number of iterations
for the bootstrapping technique, the function to be applied to the sample, and the alpha value for the probability statistic.
}
\examples{
set.seed(39)
sam<-rnorm(25,mean=25,sd=10)
myboot(x=sam,fun="mean",iter=10000, xlab="mean",col="Purple")

set.seed(40)
sam<-rgamma(30,shape=2,scale=3)
myboot(x=sam,fun="var",iter=10000,xlab="variance", col="Blue",alpha=0.20)

}
