% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{Bootstrap Simulation}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{samples used by the function}

\item{fun}{function applied by the matrix}

\item{alpha}{variable  to form confidence interval}

\item{cx}{variable for the numeric charecter expansion factor}

\item{...}{extra parameters used by the histogram}
}
\value{
a histogram showing the results of the simulation
}
\description{
Bootstrap Simulation
}
\examples{
\dontrun{
myboot(iter = 10, x=sam, alpha=0.20)
}
}
