% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Confidence Intervals}
\usage{
myci(x)
}
\arguments{
\item{x}{vector containing a sample}
}
\value{
A list containing the 95% confidence interval for ğœ‡
}
\description{
Confidence Intervals
}
\examples{
x=c(4.9769, 4.9722, 4.999, 4.9925, 4.9686, 5.0662, 4.9239, 4.9781)
myci(x)
}
