% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Central Limit Theorem Function for Poisson Distribution}
\usage{
mycltp(n, iter, lambda, ...)
}
\arguments{
\item{n}{number of values per iteration}

\item{iter}{number of iterations}

\item{lambda}{lambda value for poisson distribution}

\item{...}{additional labels for plot}
}
\value{
histogram of sampled mean, theoretical normal curve, barplot of sampled y,
and plot showing probability function for Poisson
}
\description{
Takes in the number of values per iteration, number of iterations,
and the lambda value for the generated poisson distribution and returns graphs
showing information for the Poisson distribution that is generated from the input values.
}
\examples{
mycltp(n=20,iter=10000,lambda=10)
mycltp(n=2,iter=10000,lambda=4)
mycltp(n=5,iter=10000,lambda=10)
mycltp(n=5,iter=10000,lambda=4)

}
