% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Poisson function}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{- sample size}

\item{iter}{- # of iterations}

\item{lambda}{}
}
\value{
histogram of the mean on the sample
}
\description{
Poisson function
}
\examples{
c = mycltp(10,20,10000)
}
