% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Poisson function creating a histogram of the sample mean , barplot of the sampled y and probability function for the Poisson}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{- sample size}

\item{iter}{- # of iterations}

\item{lambda}{}
}
\value{
histogram of the mean on the sample, barplot of the sampled y and probability function for the Poisson
}
\description{
Poisson function creating a histogram of the sample mean , barplot of the sampled y and probability function for the Poisson
}
\examples{
x = rnorm(30,mean=10,sd=12)
isaiahpackage::myci(x)
}
