% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{@title Central Limit Theorem Poisson}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{= row}

\item{iter}{= number of iterations}

\item{lambda}{= vector of means}

\item{...}{}
}
\value{
3 plots
}
\description{
@title Central Limit Theorem Poisson
}
\examples{
mycltp(n,inter=10000, lambda=10)
}
