% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{HYPERGEOMETRIC SIMULATION}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{Number of iterations}

\item{N}{Size}

\item{r}{1's occurences}

\item{n}{Sample size}
}
\value{
Outputs a barplot with the hypergeometric simulation and a table with the number of successes
}
\description{
HYPERGEOMETRIC SIMULATION
}
\examples{
mhyper(inter=1000,N=25,r=15,n=10)


}
