% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Distribution Plot}
\usage{
myncurve(a, mu, sigma)
}
\arguments{
\item{a}{variable to find the probability that Y<= a}

\item{mu}{mean of the distribution}

\item{sigma}{standard deviation of the distribution}
}
\value{
A plot of a normal distribution with mean mu and sd sigma along with
the lower tail probability of a
}
\description{
Normal Distribution Plot
}
\examples{
myncurve(4, 3, 2)
}
