% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mu value}

\item{sigma}{sigma value}

\item{a}{value}
}
\value{
Plot of the Normal Curve
}
\description{
Myncurve
}
\examples{
myncurve(mu=10,sigma=10, a=5)
}
