% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{Title}
\usage{
myplot(x, df)
}
\arguments{
\item{x}{a number}

\item{df}{a data frame}
}
\value{
plot
}
\description{
Title
}
\examples{

x = a num
df = spruce.df
myplot(x, spruce.df)
}
