% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{class}
\name{RLB-class}
\alias{RLB-class}
\title{Kelas Dasar untuk Regresi Linier Berganda (RLB)}
\description{
Mendefinisikan kelas S4 bernama "RLB" untuk menyimpan data dan hasil
dari model regresi linier berganda.
}
\section{Slots}{

\describe{
\item{\code{y}}{\code{numeric} Vektor respons.}

\item{\code{X}}{\code{matrix} Matriks variabel penjelas.}

\item{\code{koef}}{\code{numeric} Vektor koefisien regresi.}
}}

