% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RLB}
\alias{RLB}
\title{Konstruktor untuk Objek RLB}
\usage{
RLB(y, ...)
}
\arguments{
\item{y}{Variabel dependen (numerik).}

\item{...}{Variabel independen (numerik), akan di-cbind.}
}
\value{
Objek kelas RLB.
}
\description{
Fungsi untuk membuat dan mengestimasi model RLB.
}
