% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RLB}
\alias{RLB}
\title{Konstruktor untuk Kelas RLB}
\usage{
RLB(y, ...)
}
\arguments{
\item{y}{\code{numeric} Vektor respons.}

\item{...}{Variabel penjelas. Dapat berupa satu atau lebih vektor numerik
atau matriks tunggal dimana setiap kolom merepresentasikan variabel penjelas.}
}
\value{
Objek dari kelas \code{RLB} yang berisi y, X, dan koefisien estimasi.
}
\description{
Membuat objek dari kelas RLB untuk model regresi linier berganda.
Fungsi ini mengestimasi koefisien regresi menggunakan optimasi numerik
untuk meminimalkan jumlah kuadrat residual.
}
\examples{
\dontrun{
set.seed(123)
n <- 100
x1_data <- rnorm(n)
x2_data <- rnorm(n)
y_data <- 5 + 2*x1_data - 3*x2_data + rnorm(n)
model_rlb <- RLB(y_data, x1_data, x2_data)
print(koef.reg(model_rlb))
}
}
