% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RLB_stepwise}
\alias{RLB_stepwise}
\title{Konstruktor untuk Kelas RLB_stepwise}
\usage{
RLB_stepwise(y, ...)
}
\arguments{
\item{y}{\code{numeric} Vektor respons.}

\item{...}{Variabel penjelas kandidat. Dapat berupa satu atau lebih vektor numerik
atau matriks tunggal dimana setiap kolom merepresentasikan variabel penjelas.}
}
\value{
Objek dari kelas \code{RLB_stepwise} yang berisi y, X (variabel terpilih),
dan koefisien estimasi.
}
\description{
Membuat objek dari kelas RLB_stepwise. Pertama, fungsi ini melakukan
seleksi variabel menggunakan metode stepwise (berdasarkan AIC). Kemudian,
model regresi linier berganda dibangun menggunakan variabel-variabel terpilih.
}
\examples{
\dontrun{
set.seed(123)
n <- 100
x1_data <- rnorm(n)
x2_data <- rnorm(n) # Variabel ini mungkin tidak signifikan
x3_data <- rnorm(n)
y_data <- 5 + 2*x1_data + 0*x2_data + 3*x3_data + rnorm(n) # x2 tidak berpengaruh

model_sw <- RLB_stepwise(y_data, x1_data, x2_data, x3_data)
summary(model_sw) # Harusnya x2 tidak masuk dalam model
plot(model_sw)
}
}
