% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpackage-package.R
\docType{package}
\name{Rpackage-package}
\alias{Rpackage}
\alias{Rpackage-package}
\title{Rpackage: Have a look into your package library}
\description{
This is a practice package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/inebont/Rpackage}
  \item Report bugs at \url{https://github.com/inebont/Rpackage/issues}
}

}
\author{
\strong{Maintainer}: Ine Bonthuis \email{inebont@uio.no} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
