% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPackage.R
\name{RPackage}
\alias{RPackage}
\title{Use the xkcd API}
\usage{
RPackage(number)
}
\arguments{
\item{number}{A scalar numeric vector corresponding to an xkcd comic number}
}
\value{
A list of 13 elements
\itemize{
\item month: A scalr character vetor
\item nume: A scalar numeric vector
\item link: A scalar character vector, may be mepty
\item year: A scalar character vector
\item news: A scalar character vector, may be empty
\item safe title: A scalar character vector
\item transcript: A scalar character vector
\item alt: A scalar character vector
\item img: A scalar character vector
\item title: A scalar character vector
\item day: A scalar character vector
}
}
\description{
The \code{RPackage} function calls the xkcd JSON API and returns
metadata about the comic in the form of a list
}
\examples{
dino_time <- xkcd(15)
str(dino_time)
}
