% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_wrapper.R
\name{RSADecrypt}
\alias{RSADecrypt}
\title{RSADecrypt}
\usage{
RSADecrypt(ciphertext, d, n)
}
\arguments{
\item{ciphertext}{\itemize{
\item The string of characters to be decrypted.
}}

\item{d}{\itemize{
\item The private key decryption component. This should be an integer value
}}

\item{n}{\itemize{
\item The key product component. This should be an integer value
}}
}
\value{
plaintext - the decrypted ciphertext as a character vector
}
\description{
RSADecrypt
}
\examples{
RSADecrypt("ͶҖĥĥă ȓărĥ®ԋ", 577, 1517)
}
