% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_wrapper.R
\name{RSAPrimeGenerate}
\alias{RSAPrimeGenerate}
\title{RSAPrimeGenerate}
\usage{
RSAPrimeGenerate(n = 4)
}
\arguments{
\item{n}{\itemize{
\item the number of bits for the prime number to have
}}
}
\value{
\itemize{
\item list containing two unique prime numbers of n bit size
}
}
\description{
RSAPrimeGenerate
}
