% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_cloud.R
\name{valid_api_call}
\alias{valid_api_call}
\title{Determine if user is providing a valid API key and video ID}
\usage{
valid_api_call(api_key, video_id)
}
\arguments{
\item{api_key}{The Youtube API key}

\item{video_id}{ID of youtube video}
}
\value{
A status code of the API call
}
\description{
Determine if user is providing a valid API key and video ID
}
